% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_contrasts.R
\name{add_contrasts}
\alias{add_contrasts}
\title{Prepare limma contrasts matrix}
\usage{
add_contrasts(DIAlist, contrasts_vector = NULL, contrasts_file = NULL)
}
\arguments{
\item{DIAlist}{A DIAlist. Must have a non-empty statistical design.}

\item{contrasts_vector}{A vector of contrasts}

\item{contrasts_file}{The path to the contrasts file listing the desired contrasts.
Must be a .csv, .tsv, or .txt file.}
}
\value{
A DIAlist with added contrasts
}
\description{
Create the contrasts matrix, for use in limma.
}
\examples{
# No examples yet

}
