% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_design.R
\name{add_design}
\alias{add_design}
\title{Prepare limma model design matrix}
\usage{
add_design(DIAlist, design_formula = NULL)
}
\arguments{
\item{DIAlist}{The DIAlist to be filtered}

\item{design_formula}{NEED TO UPDATE WITH SOME SPECIFICS ON WHATS ALLOWED}
}
\value{
A DIAlist object, with a design
}
\description{
Need to rewrite
}
\examples{
# No examples yet



}
