% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata.R
\name{add_metadata}
\alias{add_metadata}
\title{Add metadata to a DIAlist}
\usage{
add_metadata(DIAlist, metadata_file)
}
\arguments{
\item{DIAlist}{A DIAlist to which metadata will be added}

\item{metadata_file}{A metadata file, giving info on the samples.}
}
\value{
A prototype of our new S3 list type, with added metadata
}
\description{
Add a dataframe of metadata to a DIAlist object. For UAMS internal use.
}
\examples{
# No examples yet

}
