% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_pw_diffs}
\alias{all_pw_diffs}
\title{All pairwise differences of a vector}
\usage{
all_pw_diffs(vector)
}
\arguments{
\item{vector}{}
}
\value{
A vector of all pairwise differences between the elements in the vector
}
\description{
Taken from \url{https://stackoverflow.com/questions/48445003/compute-all-pairwise-differences-of-elements-in-a-vector},
uses a nice R matrix algebra trick, taking the outer product of the vectors
(in this case, the difference instead of the product). Then, just return the
lower triangle, so we don't double-count pw comparisons
}
\examples{
# No examples yet

}
