% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_analysis.R
\name{check_DA_perc}
\alias{check_DA_perc}
\title{Check percentage of DA genes}
\usage{
check_DA_perc(
  DA_outcomes_table,
  DA_warn_threshold = 0.2,
  pval_thresh,
  lfc_thresh,
  adj_method
)
}
\arguments{
\item{DA_outcomes_table}{DA results data frame. Should be the output of
\code{\link[limma:decideTests]{limma::decideTests}}, coerced to a data frame.}

\item{DA_warn_threshold}{Proportion of DA genes at which we warn user.}

\item{pval_thresh}{P-value threshold used.}

\item{lfc_thresh}{logFC threshold used.}

\item{adj_method}{P-value adjustment method used.}
}
\value{
A vector of numeric values giving the \% of significant DA proteins within
each contrast.
}
\description{
Internal utility function, used in \code{\link{extract_DA_results}} to
check if assumptions are met.
}
