% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_checks.R
\name{check_data}
\alias{check_data}
\title{Check if an object is a data.frame or matrix}
\usage{
check_data(x)
}
\arguments{
\item{x}{An object to be tested (e.g., counts, metadata, annotation).}
}
\value{
Logical. Returns \code{TRUE} if \code{x} is a \code{data.frame} or \code{matrix}, otherwise returns \code{FALSE}.
}
\description{
Internal helper function to determine whether an object is a \code{data.frame} or \code{matrix}.
}
\examples{
\dontrun{
if (interactive()) {
  check_data(data.frame(a = 1:3, b = 4:6)) # TRUE
  check_data(matrix(1:9, nrow = 3))        # TRUE
  check_data(list(a = 1, b = 2))           # FALSE
}
}
}
