% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_analysis.R
\name{check_DE_perc}
\alias{check_DE_perc}
\title{Check percentage of DE genes}
\usage{
check_DE_perc(
  DE_outcomes_table,
  DE_warn_threshold = 0.2,
  pval_thresh,
  lfc_thresh,
  adj_method
)
}
\arguments{
\item{DE_outcomes_table}{DE results dataframe. Should be the output of
\code{\link[limma:decideTests]{limma::decideTests}}, coerced to a dataframe.}

\item{DE_warn_threshold}{Proportion of DE genes at which we warn user.}

\item{pval_thresh}{P-value threshold used.}

\item{lfc_thresh}{logFC threshold used.}

\item{adj_method}{P-value adjustment method used.}
}
\value{
A vector of numeric values giving the \% of significant DE proteins within
each contrast.
}
\description{
Internal utility function, used in \code{\link{extract_DA_results}} to
check if assumptions are met.
}
