% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_checks.R
\name{check_dup}
\alias{check_dup}
\title{Identify duplicate values.}
\usage{
check_dup(x)
}
\arguments{
\item{x}{a vector of values to be tested. This parameter is typically a vector of
column names, row names, or vector of values in a particular column.}
}
\value{
The function returns a vector of duplicate values.
If no duplicates are present the function returns NULL to allow
follow-up custom error messaging.
}
\description{
General purpose function that identifies duplicate values in a vector.
}
\examples{
\dontrun{
  if(interactive()){

x <- c("A", "A", "B", "C", "D", "D", "D","E")
y <- c(1, 2, 3, "A", "B", "C")

## returns vector of duplicates identified
dup_present <- check_dup(x = x)
dup_present

## returns NULL if all unique
dup_not_present <- check_dup(x = y)
dup_not_present

  }
}
}
