% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_checks.R
\name{check_file}
\alias{check_file}
\title{check if x is a file that exists}
\usage{
check_file(x)
}
\arguments{
\item{x}{an object to be tested.}
}
\value{
if x is a character string that points to
a file that exists the function returns TRUE. if x is not a character string
of length 1 that is not empty or is not a valid file name the function
returns FALSE
}
\description{
for internal use. helper function. NEED TO UPDATE
}
\examples{

\dontrun{
if(interactive()){
# no examples yet
 }
}

}
