% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_checks.R
\name{check_int}
\alias{check_int}
\title{Identify Positive Integer Values}
\usage{
check_int(x)
}
\arguments{
\item{x}{A single value to be tested. Must be of length 1L.}
}
\value{
Logical. Returns \code{TRUE} if \code{x} is a positive integer (including 0), otherwise returns \code{FALSE}.
}
\description{
Internal helper function to check if a value is a positive whole number (including zero).
}
\examples{
\dontrun{
if (interactive()) {
  check_int(5)      # TRUE
  check_int(0)      # TRUE
  check_int(-3)     # FALSE
  check_int(4.5)    # FALSE
  check_int("ABC")  # FALSE
}
}
}
