% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_checks.R
\name{check_label_column}
\alias{check_label_column}
\title{Metadata Label Column Validator}
\usage{
check_label_column(metadata, label_column, verbose = FALSE)
}
\arguments{
\item{metadata}{A \code{data.frame} containing sample metadata, where rows
represent samples.}

\item{label_column}{The name or index of the column in \code{metadata} that contains
labels used to identify each sample. The function checks if \code{label_column}:
\itemize{
\item Exists in \code{metadata}.
\item Contains non-missing and non-blank values.
\item Has unique values.
\item Adheres to R syntax rules (a warning is issued if not).
Additionally, if \code{verbose = TRUE}, a warning is displayed when label values
exceed 15 characters to encourage better plot formatting.
}}

\item{verbose}{Logical. If \code{TRUE}, checks whether the label values exceed
15 characters. Default: \code{FALSE}.}
}
\value{
The validated column name of \code{label_column}, if all checks pass.
}
\description{
Internal function for validating the sample label column in a metadata
\code{data.frame} for a \code{DGEList} object.
}
\examples{
\dontrun{
if (interactive()) {

metadata <- data.frame(
  sample = c("C1", "C2", "T1", "T2"),
  sample_id = c("C100", "C202", "T303", "T100"),
  condition = c("Control", "Control", "Treatment", "Treatment")
)

metadata2 <- data.frame(
  sample = c("C1", "C2", "T1", "T2"),
  sample = c("C100", "C202", "T303", "T100"),
  condition = c("Control", "Control", "Treatment", "Treatment"),
  check.names = FALSE
)

metadata3 <- data.frame(
  sample = c("C2", "C2", "T1", "T2"),
  sample_id = c("C100", "C202", NA, "T100"),
  sample_name = c("sample_1", "sample_2", "", "sample_4"),
  condition = c("Control", "Control", "Treatment", "Treatment")
)

metadata4 <- data.frame(
  long_sample_column_name = c("C1", "C2", "T1", "T2"),
  sample_id = c("C100", "C202", "chemo_treatment_303", "chemo_treatment_100"),
  condition = c("Control", "Control", "Treatment", "Treatment"),
  row.names = c("C100", "C202", "T303", "T100_long_row_name")
)

check_label_column(metadata = metadata, label_column = "sample_id")
check_label_column(metadata = metadata, label_column = 2)
check_label_column(metadata = metadata, label_column = "replicates")
check_label_column(metadata = metadata2, label_column = "sample")
check_label_column(metadata = metadata3, label_column = 1)
check_label_column(metadata = metadata3, label_column = 2)
check_label_column(metadata = metadata3, label_column = "sample_name")
check_label_column(metadata = metadata4, label_column = "sample_id")

}
}


}
\keyword{internal}
