% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_checks.R
\name{check_logical}
\alias{check_logical}
\title{Check if a Value is a TRUE or FALSE Logical of Length 1}
\usage{
check_logical(x)
}
\arguments{
\item{x}{A value to be checked.}
}
\value{
Logical. Returns \code{TRUE} if \code{x} is a logical value (\code{TRUE} or \code{FALSE}) of length 1.
Returns \code{FALSE} if \code{x} is any of the following: numeric, character string, length greater than 1,
empty string, a series of blank spaces, \code{NA}, \code{NULL}, or non-logical values.
}
\description{
Internal helper function to verify if \code{x} is a logical value (TRUE or FALSE) of length 1.
}
\examples{
\dontrun{
if (interactive()) {
  check_logical(TRUE)    # TRUE
  check_logical(FALSE)   # TRUE
  check_logical(1)       # FALSE
  check_logical("TRUE")  # FALSE
  check_logical(c(TRUE, FALSE)) # FALSE
  check_logical(NA)      # FALSE
}
}
}
