% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_checks.R
\name{check_long}
\alias{check_long}
\title{Identify values above a specified character length.}
\usage{
check_long(x, char_thresh = 15, na.rm = FALSE)
}
\arguments{
\item{x}{a vector of values. This parameter is typically a vector of
column names, row names, or vector of values in a particular metadata column.}

\item{char_thresh}{numeric value greater than or equal to 0.
(numeric; >= 1).
which values in x are long.}

\item{na.rm}{logical, Should NA's be omitted?. Default: FALSE}
}
\value{
The function returns a vector of values in x that exceed the
maximum number of characters threshold. If all values are below the
char_thresh the function returns NULL. If \emph{na.rm = FALSE}
any NA and NaN values in x will be included in the returned output.
\subsection{x(no NAs) + na.rm = TRUE/FALSE + all below ==> NULL}{
}

\subsection{x(NAs) + na.rm = TRUE + all below ==> NULL}{
}

\subsection{x(NAs) + na.rm = FALSE + all below ==> c(NA, NaN)}{
}

\subsection{x(no NAs) + na.rm = TRUE/FALSE + above == > c(values)}{
}

\subsection{x(NAs) + na.rm = TRUE + above ==> c(values)}{
}

\subsection{x(NAs) + na.rm = FALSE + above ==> c(values, NA, NaN)}{
}
}
\description{
General purpose function that identifies values in a vector that exceed a
maximum number of characters in length. This function is used to identify
long sample names and group labels that may lead to sub-optimal visualizations
}
\examples{

\dontrun{

 x  <- c("Con_01", "Con_02", NA, "TAC_Treatment_01", "", Inf, NaN, "TAC_Treat_02")
 y  <- c("Con_01", "Con_02", "Treat_01", "Treat_101")
 z  <- c("Con_01", NA, "TAC_Treatment_01", "TAC_Treat_02", "", Inf)

 ## identify values with more than 15 characters
 check_long(x = x, char_thresh = 15, na.rm = FALSE)

 ## identify values with more than 8 characters
 check_long(x = y, char_thresh = 8)

 ## identify values with more than 6 characters
 check_long(x = y, char_thresh = 6)

}

}
