% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_checks.R
\name{check_metadata}
\alias{check_metadata}
\title{Metadata validator.}
\usage{
check_metadata(x, verbose = FALSE)
}
\arguments{
\item{x}{a data.frame of sample metadata. Rows are samples.}

\item{verbose}{logical. Should the row names be checked for
character length? Default: FALSE.}
}
\value{
If all checks pass the function returns the input metadata as a
data.frame (invisible).
}
\description{
Internal function for validating the structure of an input metadata
data.frame.
}
\examples{
\dontrun{
if(interactive()){

metadata <- data.frame(sample = c("C1", "C2", "T1", "T2"),
                       sample_id = c("C100", "C202", "T303", "T100"),
                       condition = c("Control", "Control", "Treatment",
                                     "Treatment"))

metadata2 <- data.frame(sample = c("C1", "C2", "T1", "T2"),
                        sample = c("C100", "C202", "T303", "T100"),
                        condition = c("Control", "Control", "Treatment",
                                      "Treatment"),
                        check.names = FALSE)

metadata3 <- data.frame(long_sample_column_name = c("C1", "C2", "T1", "T2"),
                        sample_id = c("C100", "C202", "T303", "T100"),
                        condition = c("Control", "Control", "Treatment",
                                      "Treatment"),
                        row.names = c("C100", "C202", "T303",
                                      "T100_long_row_name"))

metadata4 <- data.frame(sample_id = c("C1", "C2", "T1", "T2"),
                        sample = c("C100", "C202", "T303", "T100"),
                        condition = c("Control", "Control", "Treatment",
                                      "Treatment"),
                        row.names = c("C1", "C2", "T2", ""), check.rows = FALSE)


metadata  <- check_metadata(x = metadata, verbose = TRUE)
metadata2 <- check_metadata(x = metadata2, verbose = TRUE)
metadata3 <- check_metadata(x = metadata3, verbose = TRUE)
metadata4 <- check_metadata(x = metadata4, verbose = TRUE)

 }
}

}
\keyword{internal}
