% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_checks.R
\name{check_num}
\alias{check_num}
\title{Identify Positive Numeric Values}
\usage{
check_num(x)
}
\arguments{
\item{x}{A single value to be tested.}
}
\value{
Logical. Returns \code{TRUE} if \code{x} is a valid positive numeric value (≥ 0).
Returns \code{FALSE} for non-numeric inputs, characters, multiple values, empty strings, white spaces,
\code{NA}, \code{NULL}, \code{Inf}, or negative numbers.
}
\description{
Internal helper function to verify if \code{x} is a valid positive numeric value (≥ 0).
}
\examples{
\dontrun{
if (interactive()) {
  check_num(1.4323)   # TRUE
  check_num(0)        # TRUE
  check_num(-45)      # FALSE
  check_num(Inf)      # FALSE
  check_num("snap")   # FALSE
  check_num("  ")     # FALSE
  check_num(NA)       # FALSE
  check_num(1e6)      # TRUE
}
}
}
