% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_rows_in}
\alias{check_rows_in}
\title{Internal function that checks that the rownames of set of dataframes have matching rownames in a reference set}
\usage{
check_rows_in(obj = list(), ref_rows = c())
}
\arguments{
\item{obj}{A list of dataframes to evaluate}

\item{ref_rows}{A vector of row names to which the row names of each obj dataframe will be compared}
}
\value{
A logical, with TRUE indicating all dataframes have matching rows in the reference set, and
FALSE indicates that the one or more dataframes do not have matching rownames with the reference
}
\description{
Internal function that checks that the rownames of set of dataframes have matching rownames in a reference set
}
