% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_checks.R
\name{check_string}
\alias{check_string}
\title{Validate Single Character String}
\usage{
check_string(x)
}
\arguments{
\item{x}{A value to be checked.}
}
\value{
Logical. Returns \code{TRUE} if \code{x} is a character string of length 1.
Returns \code{FALSE} if \code{x} is numeric, a vector (length > 1L), an empty string, white spaces,
\code{NA}, \code{NULL}, \code{Inf}, etc.
}
\description{
Internal helper function to check whether \code{x} is a valid character string of length 1.
}
\examples{
\dontrun{
if (interactive()) {
  check_string("ProjectID_122522") # TRUE
  check_string(123)                # FALSE
  check_string(NA)                 # FALSE
  check_string(NULL)               # FALSE
  check_string(c("oh", "snap"))    # FALSE
  check_string("     ")            # FALSE
  check_string("")                 # FALSE
  check_string(Inf)                # FALSE
}
}
}
