% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colorBatch}
\alias{colorBatch}
\title{Get colors for batches}
\usage{
colorBatch(batch)
}
\arguments{
\item{batch}{A vector of batch names.}
}
\value{
A vector of colors for each unique batch
}
\description{
Used to get colors for the batches in our missing value heatmaps.
}
\examples{
# No examples yet

}
