% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colorGroup}
\alias{colorGroup}
\title{Get colors for groups}
\usage{
colorGroup(group)
}
\arguments{
\item{group}{A vector of group names.}
}
\value{
A vector of colors for each unique group
}
\description{
Used to get colors for the groups in our missing value heatmaps.
}
\examples{
# No examples yet

}
\keyword{internal}
