% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositeS_PCA.R
\name{compositeS_PCA}
\alias{compositeS_PCA}
\title{compositeS_PCA}
\usage{
compositeS_PCA(Sfile, scale = TRUE)
}
\arguments{
\item{Sfile}{Data frame containing numerical features.}

\item{scale}{A logical value indicating whether the variables should be
scaled to have unit variance before the analysis takes place. See also
(\code{\link{prcomp}}).}
}
\value{
Data frame containing single composite score (last column)
generated from principal component analysis (PCA).
}
\description{
This function performs a Principal Component Analysis (PCA)
on the numerical feature matrix and select weights that project the
feature values on the first principal component. We recommend using this
option only when insufficient training data is available.
}
\examples{
data("Testdata_1")
# drop the length column
Testdata_1 <- Testdata_1[, -6]
# score file containing spectral counts
scoredfile1 <- cPASS_m(Testdata_1)
scoredfile1 <- scoredfile1[, -1]
# apply PCA
compS <- compositeS_PCA(scoredfile1,  scale = FALSE)
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
