% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cPASS_m.R
\name{cPASS_m}
\alias{cPASS_m}
\title{cPASS_m}
\usage{
cPASS_m(datInput)
}
\arguments{
\item{datInput}{Data frame with column names:
Experiment.id, Replicate, Bait, Prey, and counts (i.e., summed spectral
counts or log2-transformed summed MS1-intensities).}
}
\value{
Data frame containing bait-prey pairs with average peptide
spectrum match (PSMs), total PSMs, ratio total PSMs, Z-score, S-score,
D-score and WD-score.
}
\description{
This function applies Comparative Proteomic Analysis Software
Suite (CompPASS) model to score instances (e.g., bait-prey interactions
(BPIs) in the data.frame.This scoring algorithm is based on spoke model.
}
\examples{
data("Testdata_1")
# drop the length column
Testdata_1 <- Testdata_1[, -6]
# score file containing spectral counts
scoredfile1 <- cPASS_m(Testdata_1)


# score file containing MS1-intensities
# 1. first we log transform the counts
library(dplyr)
data("Testdata_2")
Testdata_2 <- Testdata_2[, -6]
Testdata_2 <-
  Testdata_2 \%>\%
  group_by(Bait, Replicate) \%>\%
  mutate(counts = log2(counts))
# 2. then we apply cPASS
scoredfile2 <- cPASS_m(Testdata_2)
}
\references{
Huttlin,E.L. et al. (2015) The BioPlex Network: A Systematic
Exploration of the Human Interactome. Cell.
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
