% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_tables.R
\name{create_combined_results}
\alias{create_combined_results}
\title{Combine statistical results}
\usage{
create_combined_results(annotation, data, statlist)
}
\arguments{
\item{annotation}{A data frame of annotation data for each gene/protein.}

\item{data}{A data frame of normalized intensity data for each sample.}

\item{statlist}{A list of per-contrast DE results.}
}
\value{
A data frame of the combined results.
}
\description{
Internal function used to construct a data frame of combined results
}
\keyword{internal}
