% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dice_m.R
\name{Dice_m}
\alias{Dice_m}
\title{Dice_m}
\usage{
Dice_m(datInput)
}
\arguments{
\item{datInput}{Data frame with column names:
Experiment.id, Bait, and Prey.}
}
\value{
Data frame containing bait-prey pairs with Dice coefficient score
greater than zero.
}
\description{
This function applies Dice coefficient to score instances
(e.g., bait-prey interactions (BPIs) in the data.frame.The Dice
coefficient was first applied by Zhang et al., 2008
to score interactions between all identified proteins (baits and preys)
in a given AP-MS experiment. This scoring algorithm is
based on matrix model that treats bait–prey and prey–prey interactions
equally.
}
\examples{
data("Testdata_1")
Testdata_1 <- Testdata_1[, c(1,3:4)]
scoredfile1 <- Dice_m(Testdata_1)
}
\references{
Zhang,B. et al. (2008) From pull-down data to protein
interaction networks and complexes with biological
relevance. Bioinformatics.
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
