% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomicsDIA-package.R
\name{dummy_statmod}
\alias{dummy_statmod}
\title{Dummy function for statmod import}
\usage{
dummy_statmod()
}
\description{
The limma functions we use need to have statmod installed, but limma
only suggests statmod, it doesn't depend on it. So, users can get halfway
through our pipeline, then have to install statmod to run a core function.
This also causes issues with running vignettes. So, we'll have a dummy
function that includes a statmod import.
}
