% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_plotting.R
\name{eval_pn_metric_for_plot}
\alias{eval_pn_metric_for_plot}
\title{Evaluate a normalization metric}
\usage{
eval_pn_metric_for_plot(
  normList,
  grouping,
  metric = c("PCV", "PMAD", "PEV", "COR", "log2ratio")
)
}
\arguments{
\item{normList}{A named list of normalized data matrices. Generally, the "normList"
slot of the list that is output by .}

\item{grouping}{A character or factor vector, listing the group(s) the samples
belong to.}

\item{metric}{The normalization metric to calculate. Can be "PCV",
"PMAD", "PEV", or "COR". See \code{\link{norm_metrics}}.}
}
\value{
A data frame containing the selected metric,
ready to be used for plotting.
}
\description{
Applies a metric (see \code{\link{norm_metrics}}) across a list of normalized
data matrices and outputs the results in a data frame for plotting.
}
\examples{
# No examples yet

}
\seealso{
\code{\link{norm_metrics}}
}
