% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_contrasts.R
\name{extract_contrast_groups}
\alias{extract_contrast_groups}
\title{Extract groups from a vector of contrasts}
\usage{
extract_contrast_groups(contrast.vec)
}
\arguments{
\item{contrast.vec}{A vector of contrasts, most likely from reading in
a contrasts text file}
}
\value{
A named list of length equal to the number of contrasts. Each element
in the list is a character vector of the groups involved in that contrast,
and the names of the list give the contrast names.
}
\description{
Extract groups from a vector of contrasts
}
\examples{
# No examples yet

}
