% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_data_import.R
\name{extract_data}
\alias{extract_data}
\title{Loading Maxquant data}
\usage{
extract_data(
  file = NULL,
  sampleIDs = NULL,
  min.prob = 0.75,
  pipe = c("DIA", "TMT", "phosphoTMT", "LF"),
  enrich = c("protein", "phospho")
)
}
\arguments{
\item{file}{The file of Maxquant data to be imported. Format and filetype
vary depending on the type of data to be analyzed.}

\item{sampleIDs}{A list of sampleIDs to analyze. Default is NULL, in which
case all sample IDs are analyzed. These are determined from column names.}

\item{min.prob}{Minimum probability, for use in local probability filter of
phosphoTMT data. Unused for other pipelines. Defauly is 0.75.}

\item{pipe}{Which analysis pipeline are you running? Options are "DIA", "TMT",
"phosphoTMT", and "LF". Output varies slightly based on the pupeline you
choose.}

\item{enrich}{Another aspect of pipeline? Options are "protein" and "phospho".
Seems semi-redundant?: DIA, TMT, and LF seem to always be "protein" in the
code. PhosphoTMT can be either protein or phospho, though that seems to vary
across subfunctions a little?}
}
\value{
A list. Items in list vary depending on pipeline used.
}
\description{
A function to extract and load MaxQuant data. Works across multiple
data types (DIA, TMT, etc.), and does slightly different things for each one.
The \code{extract_data} function is the main function, and it calls a number
of other subfunctions to process and prepare the data. See the links to
subfunctions for more info.
}
\details{
Subfunctions (DIA): \code{\link{import_data}}, \code{\link{remove_contaminants}},
\code{\link{extract_protein_data}}, \code{\link{make_log}}.
}
\examples{
# No examples yet

}
