% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_analysis.R
\name{extract_DE_results}
\alias{extract_DE_results}
\title{Extract differential expression results from a model fit}
\usage{
extract_DE_results(
  DIAlist,
  pval_thresh = 0.055,
  lfc_thresh = 1,
  adj_method = "BH"
)
}
\arguments{
\item{DIAlist}{A DIAlist, which must contain a statistical design.}

\item{pval_thresh}{The p-value threshold used to determine significance
(significant when p < pval_thresh). Default is 0.055.}

\item{lfc_thresh}{The logFC threshold used to determine significance
(significant when |logFC| > lfc.tresh). Default is 1. LogFC are base 2.}

\item{adj_method}{The method used for adjusting P-values. Default is "BH",
for the Benjamini-Hochberg correction}
}
\value{
A DIAlist object, with differential expression results.
}
\description{
Extracts statistical results describing differential expression from the.
}
\examples{
# No examples yet
}
