% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_analysis.R
\name{extract_limma_DE_results}
\alias{extract_limma_DE_results}
\title{Extract differential expression results from a model fit}
\usage{
extract_limma_DE_results(
  limma_fit,
  pval.thresh = 0.055,
  lfc.thresh = 1,
  adj.method = "BH"
)
}
\arguments{
\item{limma_fit}{A results object from the \code{\link{fit_limma_model}}
function.}

\item{pval.thresh}{The p-value threshold used to determine significance
(significant when p < pval.thresh). Default is 0.055.}

\item{lfc.thresh}{The logFC threshold used to determine significance
(significant when |logFC| > lfc.tresh). Default is 1. LogFC are base 2.}

\item{adj.method}{The method used for adjusting P-values. Default is "BH",
for the Benjamini-Hochberg correction}
}
\value{
A list with five slots \enumerate{
\item "stats_by_contrast"- A list with length equal to the number of contrasts.
Each element in the list is a data frame giving the DE results for that contrast.
\item "data"- A data frame of the data that were input into the limma DE model.
\item "pval.thresh"- The p-value threshold used to determine significance.
\item "lfc.thresh"- The logFC threshold used to determine significance.
\item "adj.method"- The p-value adjustment method used.
}
}
\description{
Extracts statistical results describing differential expression from a results
object created by \code{\link{fit_limma_model}}.
}
\examples{
# No examples yet
}
