% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_data_import.R
\name{extract_protein_data}
\alias{extract_protein_data}
\title{Extract protein data}
\usage{
extract_protein_data(data, sampleIDs = NULL, pipe, enrich)
}
\arguments{
\item{data}{Data from which to extract protein data. In the
\code{\link{extract_data}} function, this is \code{qfilterData$data}: I think this
is the dataframe of proteins with contaminants removed.}

\item{sampleIDs}{A list of sampleIDs to analyze. Default is NULL, in which
case all sample IDs are analyzed. These are determined from column names.}

\item{pipe}{Which analysis pipeline are you running? Options are "DIA", "TMT",
"phosphoTMT", and "LF". Output varies slightly based on the pupeline you
choose.}

\item{enrich}{Another aspect of pipeline? Options are "protein" and "phospho".
Seems semi-redundant?: DIA, TMT, and LF seem to always be "protein" in the
code. PhosphoTMT can be either protein or phospho, though that seems to vary
across subfunctions a little?}
}
\value{
A list. Items in list vary depending on the pipeline used.
}
\description{
Third subfunction called by \code{\link{extract_data}}. Processes the raw
protein annotation strings to extract standard gene names,
accession numbers, and Ids.
}
\examples{
# No examples yet
}
