% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_data_import.R
\name{extract_sampleIDs}
\alias{extract_sampleIDs}
\title{Extract sample IDs}
\usage{
extract_sampleIDs(colNames, sampleIDs = NULL, pipe, enrich)
}
\arguments{
\item{colNames}{A list of column names.}

\item{sampleIDs}{A list of sampleIDs to analyze. Default is NULL, in which
case all sample IDs are analyzed. These are determined from column names.}

\item{pipe}{Which analysis pipeline are you running? Options are "DIA", "TMT",
"phosphoTMT", and "LF". Output varies slightly based on the pupeline you
choose.}

\item{enrich}{Another aspect of pipeline? Options are "protein" and "phospho".
Seems semi-redundant?: DIA, TMT, and LF seem to always be "protein" in the
code. PhosphoTMT can be either protein or phospho, though that seems to vary
across subfunctions a little?}
}
\value{
A list, with a single slot. I think a vector of sample IDs.
}
\description{
A subfunction called by \code{\link{extract_protein_data}}. Honestly, not 100\%
sure what it is doing: can't tell if it is just getting the sample IDs from
column names, or whether it is subsetting dataframes down to just the columns
of sample data/intensities, and removing the metadata columns.
I think its the first one.
}
\examples{
# No examples yet
}
