% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_extension}
\alias{file_extension}
\title{Get the file extension from a file}
\usage{
file_extension(filepath)
}
\arguments{
\item{filepath}{The filepath from which to extract the extension}
}
\value{
A string of the file extension. When no extension was detected, an empty string.
}
\description{
Get the file extension from a file
}
\examples{
# No examples yet
}
