% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_process_data.R
\name{filter_data}
\alias{filter_data}
\title{Filter data}
\usage{
filter_data(data, targets, group_col = "group", min.reps = 3, min.grps = 1)
}
\arguments{
\item{data}{Raw data to be processed. In the pipeline, this is generally the
"data" slot of the list output by \code{\link{extract_data}}.}

\item{targets}{An inputs target dataframe. In the pipeline, this is generally
the subsetted targets data frame (that is, with pool or other non-desired
samples removed).}

\item{group_col}{Optional. The name of the column in the targets dataframe which
lists the experimental group names. Default is "group"}

\item{min.reps}{Optional. NEED TO FIGURE OUT}

\item{min.grps}{Optional. NEED TO FIGURE OUT}
}
\value{
NEED TO WRITE UP
}
\description{
NEED TO WRITE UP
subfunctions: make_logs
}
\examples{
# No examples yet

}
