% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_proteins.R
\name{filter_proteins_by_group}
\alias{filter_proteins_by_group}
\title{Filter protein data.}
\usage{
filter_proteins_by_group(
  DIAlist,
  min_reps = NULL,
  min_groups = NULL,
  group_col = "group"
)
}
\arguments{
\item{DIAlist}{A DIAlist to be filtered}

\item{min_reps}{The minimum number of replicates/samples within a group
that need to have non-zero intensity for a given protein/peptide for that
peptide to be considered as quantified within a group.}

\item{min_groups}{The minimum number of groups that must have at
least min_reps non-zero samples for a given protein/peptide to be retained.}

\item{group_col}{Optional. The name of the column in the targets dataframe which
lists the experimental group names. Default is "group"}
}
\value{
A filtered DIAlist
}
\description{
NEED TO REWRITE
}
\examples{
# No examples yet


}
