% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_proteins.R
\name{filter_proteins_by_proportion}
\alias{filter_proteins_by_proportion}
\title{Filter protein data by proportion.}
\usage{
filter_proteins_by_proportion(DIAlist, min_prop, grouping_column = "group")
}
\arguments{
\item{DIAlist}{A DIAlist to be filtered}

\item{min_prop}{The minimum proportion of samples within a group
in which a protein must be found for it to be retained.}

\item{grouping_column}{The name of the column in the targets dataframe which
lists the experimental group names. Default is "group"}
}
\value{
A filtered DIAlist
}
\description{
NEED TO REWRITE. Not that it rounds integer values down.
}
\examples{
# No examples yet


}
