% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_proteins.R
\name{filter_proteins_contaminants}
\alias{filter_proteins_contaminants}
\title{Remove protein contaminant rows}
\usage{
filter_proteins_contaminants(DIAlist, cols_to_search = "Protein.Name")
}
\arguments{
\item{DIAlist}{A DIAlist object to be filtered}
}
\value{
A DIAlist with contaminants filtered out
}
\description{
Not sure if this should be in internal function only, or if
it should stay public.
}
\examples{
# No examples yet

}
