% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_checks.R
\name{find_id_column}
\alias{find_id_column}
\title{Identify id column.}
\usage{
find_id_column(x, ref)
}
\arguments{
\item{x}{a data.frame. The values in each column will be queried against
the reference. x is typically a data.frame of metadata or annotation.}

\item{ref}{a character vector of reference values. The values in ref must be
unique with no NAs or blanks. ref values are typically the column names or
row names of a counts matrix (sample ids / gene ids) with x defined as
metadata or annotation.}
}
\value{
The function returns the first column name in \code{x} that contains all
reference values.
}
\description{
NEED TO UPDATE
}
\examples{

\dontrun{
if(interactive()){

metadata <- data.frame(sample = c("C1","C2","T1","T2"),
                       sample_id = c("C 100","C 202","T 303","T 100"),
                       condition = c("Control","Control","Treatment",
                                     "Treatment"))

counts <- data.frame(C.100 = c(0,10,21,3,4),
                     C.202 = c(0,0,0,2,1),
                     T.100 = c(15,33,21,55,42),
                     T.303 = c(12,22,54,33,33),
                     row.names = paste0("gene_",1:5))

annotation <- data.frame(gene_id = c("gene_2","gene_3","gene_4",
                                     "gene_1","gene_5"),
                         gene_name = c("SNP","CRC","POP","RCE","CSP"),
                         description = c("Snap","crackle","pop","rice",
                                         "crispies"))

## rsyntax error
find_id_column(x = metadata, ref = colnames(counts))
find_id_column(x = annotation, ref = row.names(counts))

  }
}

}
