% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_tables.R
\name{get_export_data}
\alias{get_export_data}
\title{Choose the best intensity matrix for export (prefer normalized per-contrast)}
\usage{
get_export_data(DAList, prefer = c("data_per_contrast", "data"))
}
\description{
Choose the best intensity matrix for export (prefer normalized per-contrast)
}
\keyword{internal}
