% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HG_m.R
\name{HG_m}
\alias{HG_m}
\title{HG_m}
\usage{
HG_m(datInput)
}
\arguments{
\item{datInput}{Data frame with column names:
Experiment.id, Replicate, Bait, Prey, and counts (i.e., summed spectral
counts or log2-transformed summed MS1-intensities) anf prey protein
length.}
}
\value{
Data frame containing bait-prey pairs with hypergeometric (HG)
probability scores.
}
\description{
This function computes the hypergeometric probability
between instances (e.g., bait-prey interactions (BPIs)) with
incorporation of NSAF. This scoring algorithm is based on matrix model
that only considers prey-prey interaction.
}
\examples{
data("Testdata_1")
# score file containing spectral counts
scoredfile1 <- HG_m(Testdata_1)


# score file containing MS1-intensities
# 1. first we log transform the counts
library(dplyr)
data("Testdata_2")
Testdata_2 <-
 Testdata_2 \%>\%
 group_by(Bait, Replicate) \%>\%
 mutate(counts = log2(counts))
# 2. then we apply HG
scoredfile2 <- HG_m(Testdata_2)
}
\references{
Guruharsha,K.G. et al. (2011) A protein complex network of
Drosophila melanogaster. Cell.
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
