% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_data_import.R
\name{import_data}
\alias{import_data}
\title{Import MaxQuant data}
\usage{
import_data(file, pipe, enrich)
}
\arguments{
\item{file}{The file of Maxquant data to be imported. Format and filetype
vary depending on the type of data to be analyzed.}

\item{pipe}{Which analysis pipeline are you running? Options are "DIA", "TMT",
"phosphoTMT", and "LF". Output varies slightly based on the pupeline you
choose.}

\item{enrich}{Another aspect of pipeline? Options are "protein" and "phospho".
Seems semi-redundant?: DIA, TMT, and LF seem to always be "protein" in the
code. PhosphoTMT can be either protein or phospho, though that seems to vary
across subfunctions a little?}
}
\value{
A list. Items in list vary depending on the pipeline used.
}
\description{
First subfunction called by \code{\link{extract_data}}. Reads in the tabular
data and does some checks for required columns.
}
\examples{
# No examples yet
}
