% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtest_impute_v9.R
\name{impute_missing_by_gtest}
\alias{impute_missing_by_gtest}
\title{Run G-test-based imputation on DAList data or contrast-specific data}
\usage{
impute_missing_by_gtest(
  DAList,
  contrast = NULL,
  grouping_column = "group",
  p_threshold = 0.01
)
}
\arguments{
\item{DAList}{A DAList object after filtering.}

\item{contrast}{Optional character string specifying contrast name (for use with data_per_contrast). If NULL, loops through all.}

\item{grouping_column}{Name of column in metadata to define group membership (default = "group").}

\item{p_threshold}{P-value threshold for triggering imputation (default = 0.05).}
}
\value{
The same DAList with imputed data written into $data or $data_per_contrast.
}
\description{
Run G-test-based imputation on DAList data or contrast-specific data
}
