% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jaccard_m.R
\name{Jaccard_m}
\alias{Jaccard_m}
\title{Jaccard_m}
\usage{
Jaccard_m(datInput)
}
\arguments{
\item{datInput}{Data frame with column names:
Experiment.id, Bait, and Prey.}
}
\value{
Data frame containing bait-prey pairs with Jaccard Index
greater than zero.
}
\description{
This function computes the Jaccard Index
for instances (e.g., bait-prey interactions (BPIs)) in
the data.frame. This scoring algorithm is based on matrix model that
treats bait–prey and prey–prey interactions equally.
}
\examples{
data("Testdata_1")
Testdata_1 <- Testdata_1[, c(1,3:4)]
scoredfile1 <- Jaccard_m(Testdata_1)
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
