% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Lou_HF.R
\docType{data}
\name{Lou_contrasts}
\alias{Lou_contrasts}
\title{Example contrasts table for Lou HF dataset}
\format{
A data frame describing example contrasts for the benchmark dataset.
}
\usage{
data(Lou_contrasts)
}
\description{
Example contrasts table for Lou HF dataset
}
\keyword{datasets}
