% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Lou_HF.R
\docType{data}
\name{Lou_HF_DIANN_uni_prot_quan}
\alias{Lou_HF_DIANN_uni_prot_quan}
\title{Example DIANN protein-level quantification for Lou HF dataset}
\format{
A data frame with one row per protein and multiple intensity columns
corresponding to the benchmark samples.
}
\usage{
data(Lou_HF_DIANN_uni_prot_quan)
}
\description{
Example DIANN protein-level quantification for Lou HF dataset
}
\keyword{datasets}
