% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Lou_HF.R
\docType{data}
\name{Lou_HF_sample_metadata}
\alias{Lou_HF_sample_metadata}
\title{Sample metadata for Lou HF dataset}
\format{
A data frame with one row per sample and several metadata columns
(e.g. sample ID, group, batch, and file name).
}
\usage{
data(Lou_HF_sample_metadata)
}
\description{
Sample metadata for Lou HF dataset
}
\keyword{datasets}
