% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_contrasts.R
\name{make_contrasts}
\alias{make_contrasts}
\title{Prepare limma contrasts matrix}
\usage{
make_contrasts(file = NULL, design)
}
\arguments{
\item{file}{The path to the contrasts file listing the desired contrasts.
Must be a .csv, .tsv, or .txt file.}

\item{design}{The limma design matrix for the model. In the pipeline, this
is usually the "design" slot of the list returned by
\code{\link{make_design}}.}
}
\value{
A list with 4 slots: \enumerate{
\item "contrasts"- The contrast matrix, for use in limma
\item "contrast.vec"- The vector of contrasts, as read from the input
contrasts file
\item "param"- A dataframe of parametes used for calling the function
\item "stats"- A dataframe giving stats (the total number of contrasts).
}
}
\description{
Create the contrasts matrix, for use in limma. Calls
\code{\link{extract_contrast_groups}} \code{\link{validate_contrasts}}
and \code{\link{make_log}} as subfunctions.
}
\examples{
# No examples yet

}
