% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_tables.R, R/limma_tables_original.R
\name{make_excel_hyperlinks}
\alias{make_excel_hyperlinks}
\title{Make hyperlinks for an Excel column}
\usage{
make_excel_hyperlinks(data, url.col, url)

make_excel_hyperlinks(data, url.col, url)
}
\arguments{
\item{data}{The data frame in which to add hyperlinks.}

\item{url.col}{The name of the column to add hyperlinks to.}

\item{url}{The url that will be prepended to the info in the column.}
}
\value{
The original data frame, now with hyperlinks in the desired column.

The original data frame, now with hyperlinks in the desired column.
}
\description{
Adds hyperlinks to a column in a data frame to be exported to an Excel file.

Internal functions to add hyperlinks to a column in a data frame to
be exported to an excel file. Implicitly, works only on one column.
}
\keyword{internal}
