% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_factor}
\alias{make_factor}
\title{Convert a column to a factor}
\usage{
make_factor(x, prefix = "X")
}
\arguments{
\item{x}{Input vector to be converted to a factor}

\item{prefix}{Optional: a prefix to put in front of numbers when
converting to a factor. Default is "X"}
}
\value{
A factor vector
}
\description{
Converts a vector/column to a factor. For numeric inputs, adds an "X"
in front of the number by default, or can supply your own prefix to use
instead.
}
\examples{
# No examples yet

}
