% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_report.R
\name{make_limma_reports}
\alias{make_limma_reports}
\title{Title}
\usage{
make_limma_reports(
  model_results = NULL,
  annotation = NULL,
  groups = NULL,
  output_dir = NULL,
  tmp_dir = "tmp",
  height = 1000,
  width = 1000,
  overwrite = F
)
}
\arguments{
\item{model_results}{TBD}

\item{annotation}{TBD}

\item{groups}{TBD}

\item{output_dir}{TBD}

\item{tmp_dir}{TBD}

\item{height}{TBD}

\item{width}{TBD}

\item{overwrite}{TBD}
}
\value{
TBD
}
\description{
Title
}
\examples{
# No examples yet
}
