% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{make_log}
\alias{make_log}
\title{Make a log file}
\usage{
make_log(param, stats, title = "", save = TRUE)
}
\arguments{
\item{param}{The param object within the \code{\link{read_DIA_data}} function
body. Slightly unclear what the requirements for this are.}

\item{stats}{The stats object within the \code{\link{read_DIA_data}} function
body. Slightly unclear what the requirements for this are.}

\item{title}{A title for the printed parameter block. Default is "".}

\item{save}{TRUE/FALSE: should the parameters be save dot a text file?}
}
\value{
A list, with the param and stats objects. If SAVE = TRUE, has side
effect of creating log files.
}
\description{
Function used to write out logs across various top-level functions
}
\examples{
# No examples yet
}
