% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_new_filename}
\alias{make_new_filename}
\title{Make new filename to avoid overwriting}
\usage{
make_new_filename(x, dir)
}
\arguments{
\item{x}{Non-unique filename we're trying to replace.}

\item{dir}{directory of the file we're trying to create}
}
\value{
A new, unique filename
}
\description{
Takes in a filename and a directory for a file that already exists. Tries to
make a new filename by adding "_01" between the filename and extension, and
continues to do so, increasing the number, until it finds a filename that does
not yet exist. Returns that filename.
}
\examples{
# No examples yet

}
