% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_report.R
\name{make_proteinorm_report}
\alias{make_proteinorm_report}
\title{Create proteinorm report}
\usage{
make_proteinorm_report(
  normList,
  groups = NULL,
  enrich = c("protein", "phospho"),
  dir = NULL,
  file = NULL,
  overwrite = FALSE,
  suppress_zoom_legend = FALSE
)
}
\arguments{
\item{normList}{A list of normalized data matrices. Generally, the "normList"
slot of the list that is output by \code{\link{process_data}}.}

\item{groups}{Optional, a vector describing what experimental or treatment
group each sample belongs to. Generally, the "group" column in the targets
data frame output by \code{\link{process_data}}. If not supplied, will
warn the user and treat all samples as belonging to the same group.}

\item{enrich}{What type of analysis is this for? Options are
"protein" and "phospho". Only used for making output dir name when one isn't
supplied.}

\item{dir}{The directory in which to save the report. If not provided,
will default to either "protein_analysis/01_quality_control" or
"phospho_analysis/01_quality_control" within the current working
directory, depending on the enrich argument.}

\item{file}{The file name of the report to be saved. Must end in .pdf. Will
default to "proteiNorm_Report.pdf" if no filename is provided.}

\item{overwrite}{Should report file be overwritten if it already exists?
Default is FALSE.}

\item{suppress_zoom_legend}{Should the legend be removed from the zoomed
log2ratio plot? Default is FALSE}
}
\value{
Invisibly, the filename of the created report
}
\description{
Creates and saves as a PDF report a variety of plots which give
information about the performance of different normalization metrics.
}
\examples{
# No examples yet

}
\seealso{
\code{\link{norm_metrics}},
\code{\link{eval_pn_metric_for_plot}},
\code{\link{pn_plots_generic}},
\code{\link{pn_plots}},
}
