% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_targets.R
\name{make_targets}
\alias{make_targets}
\title{Make a targets file}
\usage{
make_targets(input_file = NULL, sample_IDs)
}
\arguments{
\item{input_file}{Optional: a metadata file, giving info on the samples.}

\item{sample_IDs}{A character vector of sample IDs, from which the targets
file is constructed. Generally, these are the column names of the
data slot in the object returned by \code{\link{read_DIA_data}}, e.g.,
\code{colnames(ext$data)}. Currently, these sample_IDs have a required format:
"Pool" samples must include the regex pattern \code{[P/p]ool_[[:digit:]]} in
the column name, and samples must contain the regex pattern
\code{[S/s]ample_[[:digit:]]} in the column name.}
}
\value{
It depends:
\itemize{
\item If a metadata file is supplied and could be successfully matched
against the input column names, a full targets dataframe is returned.
\item If a metadata file is supplied but matching against column names
was unsuccessful, returns, with a warning, a list with two slots:
(1) the dataframe of assembled targets, and (2) the dataframe of
imported metadata. These must be manually combined before further
analysis.
\item When a metadata file is not supplied, returns, with a warning, an
incomplete targets dataframe.
}
}
\description{
Creates a dataframe of "targets". Output depends on whether a metadata file
is supplied, and whether the metadata and targets can be matched successfully
(see Value, below). Calls two subfunctions: \code{\link{import_meta}} and
\code{\link{get_dia_sample_number}}.
}
\examples{
# No examples yet

}
