% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp_missing_value_heatmaps.R
\name{missingValueHeatmap}
\alias{missingValueHeatmap}
\title{Create a missing values heatmap}
\usage{
missingValueHeatmap(
  missing,
  groups,
  batch = NULL,
  column_sort = c("cluster", "group", "batch"),
  groupColors,
  batchColors,
  sampleLabels = NULL,
  legend = FALSE
)
}
\arguments{
\item{missing}{An input matrix describing the missing data. See internals
of \code{\link{plotHeatmapsForReport}}.}

\item{column_sort}{How should the columns of the heatmap be sorted? Options
are: "cluster"- sort by similarity in missing values, "group"- sort samples
by group, "batch"- sort samples by batch.}

\item{groupColors}{A vector of colors to use for coloring samples by group}

\item{batchColors}{A vector of colors to use for coloring samples by batch.}

\item{legend}{Include a legend in the plot? Default is FALSE.}
}
\value{
A ComplexHeatmap object
}
\description{
Makes a ComplexHeatmap object showing a heatmap of missing values in the
input data.
}
\examples{
# No examples yet.

}
