% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class.R
\name{new_DAList}
\alias{new_DAList}
\title{DAList internal constructor}
\usage{
new_DAList(x = list())
}
\arguments{
\item{x}{An object to be converted into our DAList type}
}
\value{
A DAList object derived from x
}
\description{
Internal function for constructing a DAList object with no/minimal checks.
}
