% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{log2Norm}
\alias{log2Norm}
\alias{medianNorm}
\alias{meanNorm}
\alias{vsnNorm}
\alias{quantileNorm}
\alias{cycloessNorm}
\alias{rlrNorm}
\alias{giNorm}
\title{Cyclic loess normalization (optionally within groups)
Expects log2 data. If \code{groups} is given, normalizes each group independently.}
\usage{
log2Norm(dat)

medianNorm(logDat)

meanNorm(logDat)

vsnNorm(dat)

quantileNorm(logDat)

cycloessNorm(logDat, groups = NULL, method = "fast", ...)

rlrNorm(logDat)

giNorm(dat)
}
\description{
Cyclic loess normalization (optionally within groups)
Expects log2 data. If \code{groups} is given, normalizes each group independently.
}
\keyword{internal}
