% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_metrics.R
\name{norm_metrics}
\alias{norm_metrics}
\alias{PCV}
\alias{PMAD}
\alias{PEV}
\alias{COR}
\alias{log2ratio}
\title{Functions for calculating metrics for normalized data}
\usage{
PCV(data, groups)

PMAD(data, groups)

PEV(data, groups)

COR(data, groups)

log2ratio(data, groups)
}
\arguments{
\item{data}{A numeric data frame, where each row is a protein and columns
are densities. For most uses, these are probably normalized data.}

\item{groups}{A character or factor vector, listing the group(s) the samples
belong to.}
}
\value{
For PCV, PMAD, and PEV: a named vector, with length equal to the
number of groups, giving the metric for each group. For COR, a named vector
giving all within-group pairwise correlations between samples. For log2ratio,
a non-named vector giving all possible between-group log2 ratios for each
protein.
}
\description{
A set of functions that take in normalized sample data and a grouping factor
and calculate some metric of of variability, error, etc., that we can use
to evaluate normalization methods. The functions/metrics: \itemize{
\item PCV- Calculated the pooled coefficient of variation (CV, standard
deviation divided by the mean) for all proteins within a group. That is,
the average (mean) of the per-protein CVs.
\item PMAD- Calculate the pooled median absolute deviation. The MAD is the
median of the absolute deviations of each sample from the group median.
The pooled MAD is the group median of the per-protein PMADs.
\item PEV- Calculates the pooled estimate of variance for each group. Uses
the weighted average method to account for unequal sample sizes (see
\url{https://en.wikipedia.org/wiki/Pooled_variance}).
\item COR- Calculates all within-group pairwise correlations in intensity
between samples.
\item log2ratio- Calculates all possible between-group ratios of average
log2-normalized intensity for each protein.
}
}
\examples{
# No examples yet

}
