% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlap_m.R
\name{Overlap_m}
\alias{Overlap_m}
\title{Overlap_m}
\usage{
Overlap_m(datInput)
}
\arguments{
\item{datInput}{Data frame with column names:
Experiment.id, Replicate, Bait, and Prey.}
}
\value{
Data frame containing bait-prey pairs with overlap score
greather than zero.
}
\description{
This function computes the overlap similarity scores
for instances (e.g., bait-prey interactions (BPIs)) in
the data.frame. This scoring algorithm is
based on matrix model that treats bait–prey and prey–prey interactions
equally.
}
\examples{
data("Testdata_1")
Testdata_1 <- Testdata_1[, c(1,3:4)]
scoredfile1 <- Overlap_m(Testdata_1)
}
\references{
Nepusz,T. et al. (2012) Detecting overlapping protein
complexes in protein-protein interaction networks. Nat. Methods.
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
