% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{parse_protein_data}
\alias{parse_protein_data}
\title{Parse protein data}
\usage{
parse_protein_data(annotation_data)
}
\arguments{
\item{annotation_data}{Annotation data from which to parse protein data.}
}
\value{
A data frame of cleaned annotation data.
}
\description{
THIS FUNCTION FOR UAMS INTERNAL USE. This is a subfunction of \code{\link{read_DIA_data}},
used to parse protein annotation data out of the data imported from a MaxQuant file.
This function is not meant to be called by users.
}
\keyword{UAMS,}
\keyword{internal}
