% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCC_m.R
\name{PCC_m}
\alias{PCC_m}
\title{PCC_m}
\usage{
PCC_m(datInput)
}
\arguments{
\item{datInput}{Data frame with column names:
Experiment.id, Replicate, Bait, Prey, counts (i.e., summed spectral
counts or log2-transformed summed MS1-intensities) and  prey protein
length.}
}
\value{
Data frame containing bait-prey pairs with Pearson correlation
scores greather than zero.
}
\description{
This function computes the correlation
between instances (e.g., bait-prey interactions (BPIs)). This scoring
algorithm is based on matrix model that only considers prey-prey
interaction.
}
\examples{
data("Testdata_1")
# score file containing spectral counts
scoredfile1 <- PCC_m(Testdata_1)


# score file containing MS1-intensities
# 1. first we log transform the counts
library(dplyr)
data("Testdata_2")
Testdata_2 <-
 Testdata_2 \%>\%
 group_by(Bait, Replicate) \%>\%
 mutate(counts = log2(counts))
# 2. then we apply PCC
scoredfile2 <- PCC_m(Testdata_2)
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
