% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp_missing_value_heatmaps.R
\name{plotHeatmapsForReport}
\alias{plotHeatmapsForReport}
\title{Make a set of missing value heatmaps}
\usage{
plotHeatmapsForReport(
  data,
  groups,
  batch = NULL,
  sampleLabels = NULL,
  dir = ".",
  showAllProteins = FALSE,
  save = FALSE
)
}
\arguments{
\item{data}{A normalized data matrix. Generally, the "log2" slot of the normList
slot in the list output by \code{\link{process_data}}.}

\item{dir}{The directory in which to save the plot, if saving. Default is the
current working directory.}

\item{showAllProteins}{For missing data heatmaps, show all proteins (including
ones with no missing data)? Default is FALSE.}

\item{save}{Should the plot be saved (as a .png)? Default is FALSE.}
}
\value{
A list of length 4, where the first element is the missing data
matrix and the next three are the ComplexHeatmap plot objects sorted
by cluster, group, and batch, respectively.
}
\description{
Makes, and optionally saves, a set of missing value heatmaps using
\code{\link{missingValueHeatmap}}.
}
\examples{
# No examples yet.

}
