% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_report.R
\name{plotLogRatioDensity}
\alias{plotLogRatioDensity}
\title{Make a plot of the density of the log2ratio}
\usage{
plotLogRatioDensity(
  normList,
  groups,
  batch = NULL,
  zoom = FALSE,
  legend = TRUE,
  inset = 0.02,
  dir = ".",
  save = FALSE
)
}
\arguments{
\item{normList}{A list of normalized data matrices. Generally, the "normList"
slot of the list that is output by \code{\link{process_data}}.}

\item{groups}{Optional, a vector describing what experimental or treatment
group each sample belongs to. Generally, the "group" column in the targets
data frame output by \code{\link{process_data}}. If not supplied, will
warn the user and treat all samples as belonging to the same group.}

\item{batch}{Optional, a vector describing what batch each sample belongs to.
If not supplied, will treat all samples as being from the same batch.}

\item{zoom}{Should the plot cover the full range of log2ratios, or zoom
in around 0? Default is FALSE.}

\item{legend}{Include a legend in the plot? Default is TRUE.}

\item{inset}{Passed to \code{\link[graphics:legend]{graphics::legend}}, the
inset distance from the margin as a fraction of the plot region. Default is
0.02.}

\item{dir}{The directory in which to save the plot, if saving. Default is the
current working directory.}

\item{save}{Should the plot be saved (as a .png)? Default is FALSE.}
}
\value{
A list, equal in length to the input normList, where each
element of the list is the density object of the log2ratio density for a
given normalization method. See \code{\link{log2ratio}}.
}
\description{
Makes, and optionally saves, a plot showing the density of the log2ratio.
Uses \code{\link{log2ratio}} for calculating the ratio.
}
\examples{
# No examples yet

}
\seealso{
\code{\link{log2ratio}}
}
