% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_report.R
\name{plotTotInten}
\alias{plotTotInten}
\title{Make plots of total intensity for each sample}
\usage{
plotTotInten(
  normList,
  groups,
  batch = NULL,
  sampleLabels = NULL,
  dir = ".",
  save = FALSE
)
}
\arguments{
\item{normList}{A list of normalized data matrices. Generally, the "normList"
slot of the list that is output by \code{\link{process_data}}.}

\item{groups}{Optional, a vector describing what experimental or treatment
group each sample belongs to. Generally, the "group" column in the targets
data frame output by \code{\link{process_data}}. If not supplied, will
warn the user and treat all samples as belonging to the same group.}

\item{batch}{Optional, a vector describing what batch each sample belongs to.
If not supplied, will treat all samples as being from the same batch.}

\item{sampleLabels}{Optional, a set of sample labels to use. If not supplied,
defaults to using the column names in the normList.}

\item{dir}{The directory in which to save the plot, if saving. Default is the
current working directory.}

\item{save}{Should the plot be saved (as a .png)? Default is FALSE.}
}
\value{
A list, equal in length to the input normList, where each
element of the list gives total intensity for each sample for a
given normalization method.
}
\description{
Makes, and optionally saves, a set of plot showing the total intensity for
each sample across the normalization methods.
}
\examples{
# No examples yet

}
