% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_plotting.R
\name{pn_plots_generic}
\alias{pn_plots_generic}
\alias{pn_mean_plot}
\alias{pn_violin_plot}
\alias{pn_density_plot}
\title{Generic plotting functions for normalization report}
\usage{
pn_mean_plot(plotData)

pn_violin_plot(plotData)

pn_density_plot(plotData)
}
\arguments{
\item{plotData}{A dataframe of data to be plotted, created with the
\code{\link{eval_pn_metric_for_plot}} function.}
}
\value{
A ggplot object of the plot.
}
\description{
An internal set of functions that take in processed data ready for plotting and
create generic plots (e.g., means, violin plots, line plots). These generic plots
are then passed to metric-specific plotting functions (see \code{\link{pn_plots}})
to create the final plot objects.
}
\keyword{internal}
