% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_report.R
\name{prep_plot_model_data}
\alias{prep_plot_model_data}
\title{Prepare per-contrast model data for plotting}
\usage{
prep_plot_model_data(model_results, contrast)
}
\arguments{
\item{model_results}{The results slot of a DIAlist object.}

\item{contrast}{The name of the contrast for which to prep the model data}
}
\value{
A data frame of model results for the given contrast
}
\description{
Internal function used to prepare a results dataframe for both static and interactive
plots in our reports.
}
\examples{
# No examples yet
}
