% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_process_data.R
\name{process_data}
\alias{process_data}
\title{Process raw protein data}
\usage{
process_data(data, targets, group_col = "group", min.reps = 3, min.grps = 1)
}
\arguments{
\item{data}{Raw data to be processed. In the pipeline, this is generally the
"data" slot of the list output by \code{\link{extract_data}}.}

\item{targets}{An inputs target dataframe. In the pipeline, this is generally
the subsetted targets data frame (that is, with pool or other non-desired
samples removed).}

\item{group_col}{Optional. The name of the column in the targets dataframe which
lists the experimental group names. Default is "group"}

\item{min.reps}{Optional. NEED TO FIGURE OUT}

\item{min.grps}{Optional. NEED TO FIGURE OUT}
}
\value{
A list with five elements:
\enumerate{
\item "normList"- a list of length 8, where each item in the list is a
named dataframe. Names give the normalization method, and the dataframe
gives the normalized intensity data
\item "targets" - A dataframe of targets. Should be nearly the same as
the targets dataframe that was input, possibly with new rownames.
\item "filt"- The output of the \code{\link{filter_data}} function: a list
of length 5 containing filtered data and other stats.
See \code{\link{filter_data}}.
\item "param"- A dataframe giving the parameters and arguments used for
data processing
\item "stats"- A dataframe giving statistics on data processing and
filtering.
}
}
\description{
A wrapper function, which calls some fubfunctions to filter and normalize
raw data. First, calls \code{\link{filter_data}}, which does some filtering
that I do not understand yet. Then, calls \code{\link{normalize_data}}, which
runs multiple normalization methods on the the filtered data, so that later
you can pick the best normalization method. Also calls \code{\link{make_logs}}
to make some log files.
}
\examples{
# No examples yet


}
