% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_report.R
\name{proteinormMetricBoxplot}
\alias{proteinormMetricBoxplot}
\title{Make boxplot of a proteinorm metric}
\usage{
proteinormMetricBoxplot(
  normList,
  metric = c("PCV", "PMAD", "PEV", "COR"),
  groups,
  batch = NULL,
  dir = ".",
  save = FALSE
)
}
\arguments{
\item{normList}{A list of normalized data matrices. Generally, the "normList"
slot of the list that is output by \code{\link{process_data}}.}

\item{metric}{The normalization metric to calculate and plot. Can be "PCV",
"PMAD", "PEV", or "COR". See \code{\link{norm_metrics}}.}

\item{groups}{A character or factor vector, listing the group(s) the samples
belong to.}

\item{batch}{Optional, a vector describing what batch each sample belongs to.
If not supplied, will treat all samples as being from the same batch.}

\item{dir}{The directory in which to save the plot, if saving. Default is the
current working directory.}

\item{save}{Should the plot be saved (as a .png)? Default is FALSE.}
}
\value{
A list, giving the results of applying the given normalization metric
to the input normList.
}
\description{
Makes, and optionally saves, a boxplot showing the selected normalization
metric across samples. See \code{\link{norm_metrics}} for info on
the available metrics.
}
\examples{
# No examples yet

}
\seealso{
\code{\link{norm_metrics}}
}
