% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomicsDIA-package.R
\docType{package}
\name{proteoDA-package}
\alias{proteoDA}
\alias{proteoDA-package}
\title{proteoDA: Streamlined Differential Abundance Analysis of Proteomic Data}
\description{
A package for the differential proteomics analysis with DIA data. To learn more,
start with the tutorial vignette: \code{browseVignettes(package = "proteoDA")}
}
\author{
\strong{Maintainer}: Stephanie Byrum \email{sbyrum@stjude.org}

Other contributors:
\itemize{
  \item Zuo-Fei Yuan \email{zuo-fei.yuan@stjude.org} [contributor]
  \item Yingxue Fu \email{Yingxue.Fu@stjude.org} [contributor]
}

}
\keyword{internal}
