% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomicsDIA-package.R
\docType{package}
\name{proteomicsDIA-package}
\alias{proteomicsDIA}
\alias{proteomicsDIA-package}
\title{proteomicsDIA: UAMS Internal Proteomics Package}
\description{
A package for the differential proteomics anaalysis with DIA data.
}
\author{
\strong{Maintainer}: Timothy Thurman \email{tthurman@uams.edu}

Authors:
\itemize{
  \item Charity Washam \email{cwasham@uams.edu}
  \item Stephanie Byrum \email{sbyrum@uams.edu}
}

Other contributors:
\itemize{
  \item Jordan Bird \email{jtbird@uams.edu} [contributor]
}

}
\keyword{internal}
