% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting.R
\name{qc_corr_hm}
\alias{qc_corr_hm}
\title{Correlation heatmap for QC report}
\usage{
qc_corr_hm(data, groups, batch = NULL, sampleLabels = NULL)
}
\arguments{
\item{data}{A data frame of intensity data, likely normalized. Rows should be
proteins and columns should be samples.}

\item{groups}{A character or factor vector, listing the group(s) the samples
belong to.}

\item{batch}{A character or factor vector, listing the batch(es) the samples
belong to.}
}
\value{
Invisibly returns the correlation matrix created by
\code{\link[stats:cor]{stats::cor}}.
}
\description{
Plots a ComplexHeatmap object showing the pairwise correlations of intensities
across samples. Samples are grouped by similarity on the x- and y-axis, with
labels colored by group and batch.
}
\examples{
# No examples yet
}
