% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting.R
\name{qc_dendro_plot}
\alias{qc_dendro_plot}
\title{Dendrogram for QC report}
\usage{
qc_dendro_plot(
  data,
  groups = NULL,
  sample_labels = NULL,
  top_proteins = 500,
  standardize = TRUE,
  dist_metric = "euclidean",
  clust_method = "complete"
)
}
\arguments{
\item{data}{A data frame of intensity data, likely normalized. Rows should be
proteins and columns should be samples.}

\item{groups}{A character or factor vector, listing the group(s) the samples
belong to.}

\item{sample_labels}{Optional, a vector of sample labels to use. If not supplied,
defaults to using the column names in the data.}

\item{top_proteins}{The number of most variable proteins to use
for the PCA and dendrogram clustering. Default is 500.}

\item{standardize}{Should input data be standardized to a mean of 0 and std.dev of
1 before performing PCA and dendrogram clustering? If input data are
not yet standardized, should be TRUE. Default is TRUE.}

\item{dist_metric}{The metric used to define distance for dendrogram clustering.
Default is "euclidean". See \code{\link[stats:dist]{stats::hclust}} for options.}

\item{clust_method}{The agglomeration method to use for dendrogram clustering.
Default is "complete", See \code{\link[stats:hclust]{stats::hclust}} for options.}
}
\value{
A ggplot object of the plot.
}
\description{
Performs and then plots a hierarchical clustering analysis of sample intensities
across samples. Sample labels are colored according to the "groups" argument,
with colors determined by the \code{\link{colorGroup}} function. By default,
uses only the 500 most variable proteins for the analysis.
}
\keyword{internal}
