% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_dendrogram.R
\name{qc_dendrogram}
\alias{qc_dendrogram}
\title{Dendrogram for QC report}
\usage{
qc_dendrogram(
  data,
  groups = NULL,
  sample_labels = NULL,
  text.sizes = c(14, 3, 12),
  point.size = 3,
  top = 500,
  standardize = TRUE,
  dist_metric = "euclidean",
  clust_method = "complete",
  colors = NULL,
  legend.position = "right",
  title = NULL,
  subtitle = NULL,
  show.plot = TRUE
)
}
\arguments{
\item{standardize}{Should input data be standardized to a mean of 0 and std.dev of
1 before performing PCA and dendrogram clustering? If input data are
not yet standardized, should be TRUE. Default is TRUE.}

\item{dist_metric}{The metric used to define distance for dendrogram clustering.
Default is "euclidean". See \code{\link[stats:dist]{stats::dist}} for options.}

\item{clust_method}{The agglomeration method to use for dendrogram clustering.
Default is "complete", See \code{\link[stats:hclust]{stats::hclust}} for options.}
}
\value{
A ggplot object of the plot.
}
\description{
Performs and then plots a hierarchical clustering analysis of sample intensities
across samples. Sample labels are colored according to the "groups" argument,
with colors determined by the \code{\link{colorGroup}} function. By default,
uses only the 500 most variable proteins for the analysis.
}
\keyword{internal}
