% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_dendrogram.R
\name{qc_dendrogram_subgroups}
\alias{qc_dendrogram_subgroups}
\title{Dendrogram for QC Report with Subgroups}
\usage{
qc_dendrogram_subgroups(
  DAList,
  grouping_column,
  label_column = NULL,
  text.sizes = c(12, 4, 12),
  point.size = 3,
  group_color = NULL,
  top = 500,
  standardize = TRUE,
  dist_metric = "euclidean",
  clust_method = "complete",
  legend.position = "right",
  show.plot = TRUE
)
}
\arguments{
\item{DAList}{A list containing data and metadata.}

\item{grouping_column}{A character string specifying the sample metadata column to use for subgrouping (e.g., "group","batch", etc)}

\item{label_column}{An optional character string specifying the sample metadata column for sample labels.}

\item{text.sizes}{A numeric vector specifying the text sizes for the title, leaf labels, and legend text.}

\item{point.size}{A numeric vector specifying the size of the node circles.}

\item{group_color}{A character string specifying the highlight color for the subgroup.}

\item{top}{The number of the most variable proteins to use in the analysis.}

\item{standardize}{A logical value indicating whether to standardize data (e.g., mean = 0, SD= 1).}

\item{dist_metric}{A character string specifying the distance metric (e.g., "euclidean", "manhattan").}

\item{clust_method}{A character string specifying the clustering method (e.g, "complete", "ward.D2")}

\item{legend.position}{A character string specifying the legend position (e.g., "right", "left", "top", "bottom").}

\item{show.plot}{A logical value indicating whether to display the plot.}
}
\value{
A list of dendrogram plots for each subgroup.
}
\description{
Generates dendrograms for different subgroups within a dataset.
Each dendrogram highlights one subgroup while keeping others in grayscale.
}
\examples{
\dontrun{
# An example of plot
den-> qc_dendrogram_subgroups(DAList = results, 
                               grouping_column = 
den <- qc_dendrogram(DAList           = results,
                    grouping_column  = results$metadata$group,
                    label_column     = results$metadata$sample,
                    top           = nrow(results$data),
                    standardize   = TRUE,
                    dist_metric   = "euclidean",
                    clust_method  = "complete",
                    colors        = all_colors2$group,
                    point.size    = 3,
                    text.sizes    = c(14,3,9),
                    legend.position = "right",
                    title         = "",
                    show.plot     = FALSE)

}
}
