% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting.R
\name{qc_missing_hm}
\alias{qc_missing_hm}
\title{Create a missing values heatmap}
\usage{
qc_missing_hm(
  data,
  groups,
  sample_labels = NULL,
  column_sort = c("cluster", "group"),
  group_var_name = "",
  show_all_proteins = F
)
}
\arguments{
\item{data}{A data frame of intensity data, likely normalized. Rows should be
proteins and columns should be samples. Generally, a single element in the
normList slot of the object returned by .}

\item{groups}{A character or factor vector, listing the group(s) the samples
belong to.}

\item{sample_labels}{Optional, a vector of sample labels to use. If not supplied,
defaults to using the column names in the data.}

\item{column_sort}{How should the columns of the heatmap be sorted? Options
are: "cluster"- sort by similarity in missing values, "group"- sort samples
by the grouping variable.}

\item{group_var_name}{The name of the variable being used for group sorting.}

\item{show_all_proteins}{Should all proteins be shown in missing value heatmap,
of only those with missing data? Default is F (only those with missing data).}
}
\value{
A \code{\link[grid:gTree]{grid::gTree}} object of the ComplexHeatmap,
which can be plotted with \code{\link[grid:grid.draw]{grid::grid.draw}}.
}
\description{
Makes a ComplexHeatmap object showing a heatmap of missing values in the
input data.
}
\examples{
# No examples yet.

}
\keyword{internal}
