% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting.R
\name{qc_pca_plot}
\alias{qc_pca_plot}
\title{PCA plot for QC report}
\usage{
qc_pca_plot(
  data,
  groups = NULL,
  sample_labels = colnames(data),
  top_proteins = 500,
  standardize = TRUE,
  pca_axes = c(1, 2)
)
}
\arguments{
\item{data}{A data frame of intensity data, likely normalized. Rows should be
proteins and columns should be samples. Generally, a single element in the
normList slot of the object returned by .}

\item{groups}{A character or factor vector, listing the group(s) the samples
belong to.}

\item{sample_labels}{Optional, a vector of sample labels to use. If not supplied,
defaults to using the column names in the data.}

\item{top_proteins}{The number of most variable proteins to use for the analysis.
Default is 500.}

\item{standardize}{Should input data be standardized to a mean of 0 and std.dev of
1? If input data are not yet standardized, should be TRUE. Default is TRUE.}

\item{pca_axes}{A numeric vector of length 2 which lists the PC axes to plot.
Default is c(1,2), to plot the first two principal components.}
}
\value{
A ggplot object of the plot.
}
\description{
Performs and then plots a principal component analysis of sample intensities.
Samples are colored according to the groups argument, with colors
determined by the \code{\link{colorGroup}} function. By default, uses
only the 500 most variable proteins for the analysis.
}
\examples{
# No examples yet
}
\keyword{internal}
