% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_pca.R
\name{qc_pca_plot7}
\alias{qc_pca_plot7}
\title{PCA Plot}
\usage{
qc_pca_plot7(
  data,
  groups = NULL,
  sample_labels = NULL,
  text.sizes = c(14, 14, 12, 10),
  point.size = 3,
  label.size = 4,
  max.labels = 100,
  top = 500,
  standardize = TRUE,
  pca_axes = c(1, 2),
  colors = NULL,
  title = NULL,
  max_pc = NULL,
  legend.position = "right",
  show.plot = TRUE
)
}
\arguments{
\item{data}{A data frame of intensity data, likely normalized. Rows should be proteins and columns should be samples.}

\item{groups}{A character or factor vector, listing the group(s) the samples belong to. Default: NULL}

\item{sample_labels}{Optional, a vector of sample labels to use. If not supplied, defaults to using the column names in the data. Default: NULL}

\item{text.sizes}{A numeric vector specifying text sizes for different plot elements: title, axis titles, axis labels, and legend. Default: c(14, 14, 12, 10)}

\item{point.size}{Size of points in the PCA plot. Default: 3}

\item{label.size}{Size of sample labels in the plot. Default: 4}

\item{max.labels}{Maximum number of labels to show in the plot. Default: 100}

\item{top}{The number of most variable proteins to use for the PCA analysis. Default: 500}

\item{standardize}{Logical indicating whether to standardize the input data before PCA. Default: TRUE}

\item{pca_axes}{A numeric vector of length 2 specifying which PCA axes to plot. Default: c(1, 2)}

\item{colors}{A vector of colors corresponding to groups. Default: NULL}

\item{title}{Title for the plot. Default: NULL}

\item{max_pc}{Optional integer giving the maximum principal component
index to include in the scree plot produced by
\code{qc_pca_scree_plot7()}. Default: NULL (shows all PCs).}

\item{legend.position}{Position of the legend in the plot. Default: 'right'}

\item{show.plot}{Logical indicating whether to display the plot. Default: TRUE}
}
\value{
A list containing the PCA plot, the PCA analysis results, and additional information.
}
\description{
Generates PCA plots from normalized intensity data.
}
\details{
This function performs PCA on the provided data and generates a corresponding plot.
}
\examples{
\dontrun{
if(interactive()){
 # Example usage
 qc_pca_plot7(data, groups)
 }
}
}
\seealso{
\code{\link[rlang]{arg_match}}, \code{\link[cli]{cli_abort}},
\code{\link[stats]{prcomp}}, \code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{aes}},
\code{\link[ggrepel]{geom_label_repel}}, \code{\link[ggplotify]{as.ggplot}}
}
