% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_pca.R
\name{qc_pca_plot7_subgroups}
\alias{qc_pca_plot7_subgroups}
\title{PCA Plot for Subgroups}
\usage{
qc_pca_plot7_subgroups(
  DAList,
  grouping_column,
  label_column = NULL,
  group_color = NULL,
  top = 500,
  standardize = TRUE,
  pca_axes = c(1, 2),
  text.sizes = c(14, 14, 12, 10),
  point.size = 3,
  label.size = 4,
  max.labels = 100,
  legend.position = "right",
  show.plot = TRUE,
  max_pc = NULL
)
}
\arguments{
\item{DAList}{A list containing data and metadata for differential analysis. Should include a 'data' matrix and 'metadata' data frame.}

\item{grouping_column}{The name of the column in the metadata that defines the grouping for the samples.}

\item{label_column}{The name of the column in the metadata for sample labels. Default: NULL}

\item{group_color}{A color to represent the groups in the plot. Default: NULL (will use default color).}

\item{top}{The number of the most variable proteins to include in the PCA analysis. Default: 500}

\item{standardize}{Logical indicating whether to standardize the input data to have mean 0 and standard deviation 1. Default: TRUE}

\item{pca_axes}{A numeric vector of length 2 specifying which PCA axes to plot. Default: c(1, 2)}

\item{text.sizes}{A numeric vector specifying text sizes for different plot elements: title, axis titles, axis labels, and legend. Default: c(14, 14, 12, 10)}

\item{point.size}{Size of points in the PCA plot. Default: 3}

\item{label.size}{Size of sample labels in the plot. Default: 4}

\item{max.labels}{Maximum number of labels to show in the plot. Default: 100}

\item{legend.position}{Position of the legend in the plot. Default: 'right'}

\item{show.plot}{Logical indicating whether to display the plot. Default: TRUE}
}
\value{
A list containing PCA plots for each subgroup and the PCA analysis results.
}
\description{
Generates PCA plots for different subgroups based on provided metadata.
}
\details{
This function allows users to visualize the PCA results for different subgroups in the dataset, providing insights into the variance and structure of the data.
}
\examples{
\dontrun{
if(interactive()){
 # Example usage
 qc_pca_plot7_subgroups(DAList, "group", "sample_label")
 }
}
}
