% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_pca.R
\name{qc_pca_scree_plot7}
\alias{qc_pca_scree_plot7}
\title{Scree Plot for PCA}
\usage{
qc_pca_scree_plot7(pca, max_pc = NULL, text.sizes = c(16, 14, 14))
}
\arguments{
\item{pca}{A prcomp object resulting from PCA analysis.}

\item{max_pc}{The maximum number of PCs to display in the scree plot. Default: NULL (shows all PCs)}

\item{text.sizes}{A numeric vector specifying text sizes for different plot elements: title, axis titles, axis labels. Default: c(16, 14, 14)}
}
\value{
A list containing the scree plot and the data used for plotting.
}
\description{
Generates a scree plot showing the proportion of variance for each principal component.
}
\details{
This function visualizes the variance explained by each principal component in the PCA analysis.
}
\examples{
\dontrun{
if(interactive()){
 # Example usage
 qc_pca_scree_plot7(pca)
 }
}
}
\seealso{
\code{\link[cli]{cli_abort}}
}
