% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_violin.R
\name{qc_violin}
\alias{qc_violin}
\title{Generate a Violin Plot for Intensity Data}
\usage{
qc_violin(
  data,
  groups = NULL,
  sample_labels = NULL,
  title = "",
  text.sizes = c(12, 10, 10, 10),
  legend.position = "right",
  color_palette = NULL
)
}
\arguments{
\item{data}{A matrix or data frame containing intensity values.}

\item{groups}{A vector (numeric, integer, character, or factor) indicating class membership. Defaults to NULL.}

\item{sample_labels}{A character vector specifying sample names to display on the plot. Defaults to column names of \code{data}.}

\item{title}{A character string specifying the plot title. Defaults to an empty string.}

\item{text.sizes}{A numeric vector of length 4 specifying font sizes for title, x-axis, y-axis, and legend text, respectively.}

\item{legend.position}{A character string specifying legend position (e.g., "right", "left", "none"). Defaults to "right".}

\item{color_palette}{A character vector specifying a colorblind-friendly palette. If NULL, a default base palette is used based on RColorBrewer}
}
\value{
A ggplot object representing the violin plot.
}
\description{
Generate a Violin Plot for Intensity Data
}
\examples{
\dontrun {
# example of violin plot colored by groups 
violin1 <- qc_violin(data            = results$data,
                     groups          = results$metadata$group,
                     sample_labels   = results$metadata$sample,
                     title           = "",
                     text.sizes      = c(12, 10, 10, 10),
                     legend.position = "right",
                     color_palette   = c("#66c2A5","#8DA0CB", "#FC8D62")  
                     )


}


}
