% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_commas}
\alias{remove_commas}
\title{Removes commas in character representation of a number}
\usage{
remove_commas(x)
}
\arguments{
\item{x}{Character string of a number, with commas in it.}
}
\value{
Integer
}
\description{
Removes commas in character representation of a number
}
\keyword{internal}
