% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_contaminants.R
\name{remove_contaminants}
\alias{remove_contaminants}
\title{Remove protein contaminant rows}
\usage{
remove_contaminants(data)
}
\arguments{
\item{data}{Raw Maxquant intensity data, from which contaminants need to be
removed. Within the \code{\link{read_DIA_data}} function, this is "diaData",
which is the data object from the \code{\link{read_maxquant_delim}} function that has
gone through additional processing within the body of the
\code{\link{read_DIA_data}} function}
}
\value{
A list of data with contaminants removed, with three elements:
\enumerate{
\item "data", a data frame with the contaminants removed. Rows are
proteins, columns include both protein info and individual intensities.
\item "num_contam", an integer giving the number of contaminant rows removed
\item "num_qfilter", an integer giving the number of retained rows
(ie, nrow() of the "data" slot).
}
}
\description{
The second subfunction called by \code{\link{read_DIA_data}}. Removes
contaminant rows from the data.
}
\examples{
# No examples yet

}
