% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rowMads}
\alias{rowMads}
\title{Calculate per-row MADs of a numeric array}
\usage{
rowMads(x, ...)
}
\arguments{
\item{x}{The array for which to calculate per-row MADs}

\item{...}{Additional arguments to be passed to internal functions.
Meant for na.rm.}
}
\value{
A numeric vector of appropriate length, named if input was named, with
per-row MADs
}
\description{
Have done testing, gives same results as the matrixStats::rowMads() function
it replaces, though it is much slower (the matrxiStats version uses C code).
}
\examples{
# No examples yet.
}
