% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rowSds}
\alias{rowSds}
\title{Calculate per-row standard deviations of a numeric array}
\usage{
rowSds(x, ...)
}
\arguments{
\item{x}{The array for which to calculate per-row Sds}

\item{...}{Additional arguments to be passed to internal functions.
Meant for na.rm.}
}
\value{
A numeric vector of appropriate length, named if input was named, with
per-row standard deviations
}
\description{
Have done testing, gives same results as the genefilter::rowSds() function.
}
\examples{
# No examples yet.
}
