% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rowVars}
\alias{rowVars}
\title{Calculate per-row variances of a numeric array}
\usage{
rowVars(x, ...)
}
\arguments{
\item{x}{The array for which to calculate per-row variances}

\item{...}{Additional arguments to be passed to internal functions.
Meant for na.rm}
}
\value{
A numeric vector of appropriate length, named if input was named, with
per-row variances.
}
\description{
Have done testing, gives same results as the genefilter::rowVars() function.
}
\examples{
# No examples yet.
}
\keyword{internal}
