% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_APMS.R
\name{score_APMS}
\alias{score_APMS}
\title{Scoring Affinity Purification Mass Spectrometry (AP-MS) Datasets
via spoke and matrix models.}
\usage{
score_APMS(
  datInput,
  cPASS_m = TRUE,
  MiST_m = TRUE,
  HG_m = FALSE,
  WM_m = FALSE,
  PCC_m = FALSE,
  Dice_m = FALSE,
  Jaccard_m = FALSE,
  Overlap_m = FALSE,
  Zscore_m = FALSE,
  iter = 3
)
}
\arguments{
\item{datInput}{Data frame with column names:
Experiment.id, Replicate, Bait, Prey, counts (i.e., summed spectral
counts or log2-transformed summed MS1-intensities), and
prey protein length.}

\item{cPASS_m}{If TRUE, scores the associations
using Comparative Proteomic Analysis Software Suite (CompPASS) model.
This algorithm is based on spoke model.}

\item{MiST_m}{If TRUE, scores the associations using
Mass spectrometry interaction STatistics (MiST) model.
This algorithm is based on spoke model.}

\item{HG_m}{If TRUE, scores the associations
via hypergeometric probability with
incorporation of NSAF. This algorithm is based on matrix model.}

\item{WM_m}{If TRUE, scores the associations
using the weighted matrix model.This algorithm is based on matrix model.}

\item{PCC_m}{If TRUE, scores the associations
using PCC score. This algorithm is based on matrix model.}

\item{Dice_m}{If TRUE, scores the associations
using Dice metric.This algorithm is based on matrix model.}

\item{Jaccard_m}{If TRUE, scores the associations
using Jaccard metric.This algorithm is based on matrix model.}

\item{Overlap_m}{If TRUE, scores the associations
using Overlap score.This algorithm is based on matrix model.}

\item{Zscore_m}{If TRUE, scores the associations
using Z-score.This algorithm is based on matrix model.}

\item{iter}{Number of bootstraps to execute to estimate z scores.
Defaults to 3.}
}
\value{
A data frame containing the calculated features for Bait-Prey
or Prey-Prey associations.
}
\description{
This function computes Bait-prey and Prey-Prey associations
via nine metrics: (Comparative Proteomic Analysis Software Suite (CompPASS)
(\code{\link{cPASS_m}}),
Mass spectrometry interaction STatistics (\code{\link{MiST_m}}),
Hypergeometric probability (\code{\link{HG_m}}),
Weighted matrix model (\code{\link{WM_m}}),
Dice (\code{\link{Dice_m}}), Jaccard (\code{\link{Jaccard_m}}),
ovelrap (\code{\link{Overlap_m}}),Z_score
(\code{\link{Zscore_m}}), and Pearson correlation coefficients
(\code{\link{PCC_m}}).By default ComPASS and MiST scoring is enabled,
optionally, one or more of these can be enabled or disabled.
}
\details{
score_APMS
}
\examples{
# score file containing spectral counts
data(Testdata_1)
scoredfile1 <-  score_APMS(Testdata_1)

# score file containing MS1-intensities
# 1. first we log transform the counts
library(dplyr)
data("Testdata_2")
Testdata_2 <-
  Testdata_2 \%>\%
  group_by(Bait, Replicate) \%>\%
  mutate(counts = log2(counts))
# 2. then we apply score_APMS
scoredfile2 <- score_APMS(Testdata_2)
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
