% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_report.R
\name{static_MD_plot}
\alias{static_MD_plot}
\title{Make a DE MD plot.}
\usage{
static_MD_plot(data, lfc.thresh, contrast, pval.type)
}
\arguments{
\item{data}{Per-contrast DE results to be plotted, as prepared by
\code{\link{prep_plot_model_data}}.}

\item{lfc.thresh}{The logFC threshold used to determine significance
(significant when |logFC| > lfc.tresh). LogFC are base 2.}

\item{contrast}{The contrast being plotted. Used for generating the plot title.}

\item{pval.type}{The type of p-value to plot. Can be "raw" or "adjusted".}
}
\value{
A ggplot object
}
\description{
Internal function for plotting static versions of MD plots.
}
\examples{
# No examples yet
}
