% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Static_Volcano.R, R/limma_report.R
\name{static_volcano_plot}
\alias{static_volcano_plot}
\title{Make a DE Volcano plot.}
\usage{
static_volcano_plot(
  data,
  lfc_thresh,
  pval_thresh,
  contrast,
  pval_type,
  control_proteins = NULL,
  anno = NULL,
  highlight_by = "uniprot_id"
)

static_volcano_plot(
  data,
  lfc_thresh,
  pval_thresh,
  contrast,
  pval_type,
  control_proteins = NULL,
  anno = NULL,
  highlight_by = "uniprot_id"
)
}
\arguments{
\item{data}{Per-contrast DE results to be plotted, as prepared by
\code{\link{prep_plot_model_data}}.}

\item{lfc_thresh}{The logFC threshold used to determine significance
(significant when |logFC| > lfc.tresh). LogFC are base 2.}

\item{pval_thresh}{The p-value threshold used to determine significance
(significant when p < pval_thresh).}

\item{contrast}{The contrast being plotted. Used for generating the plot title.}

\item{pval_type}{The type of p-value to plot. Can be "raw" or "adjusted".}

\item{control_proteins}{Character vector of protein IDs/names to highlight.}

\item{anno}{Data frame of annotation info (must include highlight_by column).}

\item{highlight_by}{Column name in \code{anno} used to match \code{control_proteins}.}
}
\value{
A ggplot object.

A ggplot object.
}
\description{
Internal function for plotting static versions of volcano plots.

Internal function for plotting static versions of volcano plots.
}
\keyword{"GN"}
\keyword{"Gene.names"}
\keyword{"KO_vs_WT")}
\keyword{"KO_vs_WT",}
\keyword{"MDM2"),}
\keyword{"TP53"),}
\keyword{"adjusted",}
\keyword{"uniprot_id"}
\keyword{#}
\keyword{#'}
\keyword{)}
\keyword{0.05,}
\keyword{1,}
\keyword{300)}
\keyword{5)}
\keyword{5,}
\keyword{6,}
\keyword{<-}
\keyword{=}
\keyword{@examples}
\keyword{\dontrun{}
\keyword{anno}
\keyword{c("P04637",}
\keyword{c("TP53",}
\keyword{contrast}
\keyword{control_proteins}
\keyword{dalist$anno}
\keyword{dalist$anno,}
\keyword{dalist$results,}
\keyword{data}
\keyword{dpi}
\keyword{gene}
\keyword{ggsave("volcano_KO_vs_WT.pdf",}
\keyword{ggsave("volcano_KO_vs_WT.png",}
\keyword{height}
\keyword{highlight_by}
\keyword{internal}
\keyword{lfc_thresh}
\keyword{name}
\keyword{or}
\keyword{p}
\keyword{p,}
\keyword{plot_data}
\keyword{plot_data,}
\keyword{prep_plot_model_data(model_results}
\keyword{pval_thresh}
\keyword{pval_type}
\keyword{static_volcano_plot(}
\keyword{uniprot_id}
\keyword{width}
\keyword{}}
