% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_targets.R
\name{subset_targets}
\alias{subset_targets}
\title{Filter samples from targets file}
\usage{
subset_targets(targets, filter_list, ignore.case = TRUE)
}
\arguments{
\item{targets}{The input targets dataframe to be filtered}

\item{filter_list}{A named list describing how to filter targets dataframe.
The name for each element of the list gives a column to be filters, and
each element of the list should be a character vector of strings to search for
in the given column. See examples.}

\item{ignore.case}{Should the case of the strings in rm.vals be ignored? Is
TRUE by default, such that case does not have to match. If FALSE, case must
match for rows to be removed.}
}
\value{
A list with three elements:
\enumerate{
\item "targets", a dataframe of subsetted targets
\item "stats", a dataframe with statistics on filtering
\item "param", a dataframe giving the parameters used for filtering
}
}
\description{
Removes unwanted samples (rows) from the targets data frame This function is
basically \code{\link[base]{subset}} with extra functionality. It removes
rows if they contain certain values in a given column. By default, ignores
the case of strings in rm.val. Calls \code{\link{make_log}} as a subfunction.
}
\examples{
\dontrun{
sub_targets <- subset_targets(targets = target,
                              filter_list = list(group = "pool"
                                                 sample = c("sampleA", "sampleB")))
}

}
