% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_tables.R
\name{summarize_contrast_DA}
\alias{summarize_contrast_DA}
\title{Summarize the number of DA proteins in a contrast}
\usage{
summarize_contrast_DA(contrast_name, contrast_res_list)
}
\arguments{
\item{contrast_name}{The name of the contrast to summarize.}

\item{contrast_res_list}{A list of per-contrast DA results.}
}
\value{
A data frame summarizing differential abundance for the given contrast.
}
\description{
Internal function to summarize the number of DE genes/proteins for a given
contrast.
}
\keyword{internal}
