% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_tables.R
\name{summarize_contrast_DE}
\alias{summarize_contrast_DE}
\title{Summarize the number of DE proteins in a contrast}
\usage{
summarize_contrast_DE(contrast_name, contrast_res_list)
}
\arguments{
\item{contrast_name}{The name of the contrast to summarize}

\item{contrast_res_list}{A list of per-contrast DE results}
}
\value{
A data frame summarizing DE for the given contrast
}
\description{
Internal function to summarize the number of DE genes/proteins for a given
contrast.
}
\examples{
# No examples yet
}
