% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testdata_1.R
\docType{data}
\name{Testdata_1}
\alias{Testdata_1}
\title{Test data 1 for AP-MS scoring}
\source{
\url{https://uregina.ca/~babu200m/index.html}
}
\usage{
data(Testdata_1)
}
\description{
Data input including summed spectral counts
\itemize{
  \item Experiment ID = sample identifier per AP-MS run
  \item Replicate = biological replicates
  \item Bait = a unique identifier (i.e., UniProt accession code) per Bait
  \item Prey = a unique identifier (i.e., UniProt accession code) per Prey
  \item counts = Summed spectral counts per proteins
  \item Length = Prey length
}
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
\keyword{datasets}
