% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testdata_2.R
\docType{data}
\name{Testdata_2}
\alias{Testdata_2}
\title{Test data 2 for AP-MS scoring}
\source{
\url{https://uregina.ca/~babu200m/index.html}
}
\usage{
data(Testdata_2)
}
\description{
Data input including summed MS1-intensities per protein
\itemize{
  \item Experiment ID = sample identifier per AP-MS run
  \item Replicate = biological replicates
  \item Bait = a unique identifier (i.e., UniProt accession code) per Bait
  \item Prey = a unique identifier (i.e., UniProt accession code) per Prey
  \item counts = Summed MS1-intensities per protein
  \item Length = Prey length
}
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
\keyword{datasets}
