% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useless_function.R
\name{useless_function}
\alias{useless_function}
\title{Do something useless}
\usage{
useless_function(...)
}
\arguments{
\item{...}{Any argument you want, named or unnamed.}
}
\value{
invisible(NULL)
}
\description{
A function that doesn't do anything (besides serving as an example about how to test and document functions).
}
\examples{
x <- useless_function(1234)

}
