% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class.R
\name{validate_DAList}
\alias{validate_DAList}
\title{DAList validator}
\usage{
validate_DAList(x)
}
\description{
Internal function for validating a DAList object.
Tolerant of arbitrary result labels (custom contrast names).
}
\keyword{internal}
