% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class.R
\name{validate_DIAlist}
\alias{validate_DIAlist}
\title{DIAlist internal constructor}
\usage{
validate_DIAlist(x)
}
\arguments{
\item{x}{An object to be converted into our DIAlist type}
}
\value{
A prototype of our new S3 list type (currently just a list).
}
\description{
Internal function for constructing a DIAlist object with no/minimal checks.
}
\examples{
# No examples yet

}
