% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_filename}
\alias{validate_filename}
\title{Validate filename}
\usage{
validate_filename(filename, allowed_exts, check_space = T)
}
\arguments{
\item{filename}{File name to check}

\item{allowed_exts}{Character vector of allowable extensions, with no period}

\item{check_space}{Check for whitespace in file name?}
}
\value{
If filename checks out, returns invisible TRUE. Otherwise, throws
error.
}
\description{
Internal function used to check constraints on user-defined filenames. By
default checkes whether there are spaces in the filename, and whether the
extension is in the list of allowed extensions.
}
\examples{
# No examples yet
}
