% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_design.R
\name{validate_formula}
\alias{validate_formula}
\title{Validate a statistical formula}
\usage{
validate_formula(design_formula)
}
\arguments{
\item{design_formula}{A string or expression which represents a statistical formula}
}
\value{
If all checks pass, a formula object of the design_formula
}
\description{
An internal function to validate
}
\examples{
# No examples yet


}
