% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WM_m.R
\name{WM_m}
\alias{WM_m}
\title{WM_m}
\usage{
WM_m(datInput)
}
\arguments{
\item{datInput}{Data frame with column names:
Experiment.id, Bait, and Prey.}
}
\value{
Data frame containing bait-prey pairs with
k (i.e.,number of co-purifications) greather or equal to 1
& logHG (i.e., -1*log(P-val) of the hypergeometric test).
}
\description{
This function computes the weighted matrix model for
instances (e.g., bait-prey interactions (BPIs)) in the data.frame.
The output of the weighted matrix model includes the number of experiments
for which the pair of proteins is co-purified (i.e., k) and
-1*log(P-value) of the hypergeometric test (i.e., logHG) given the
experimental overlap value, each protein's total number of observed
experiments, and the total number of experiments (Drew et al., 2017).
This scoring algorithm is based on matrix model that treats
bait–prey and prey–prey interactions equally.
}
\examples{
data("Testdata_1")
Testdata_1 <- Testdata_1[, c(1,3:4)]
scoredfile1 <- WM_m(Testdata_1)
}
\references{
Drew,K. et al. (2017) Integration of over 9,000 mass
spectrometry experiments builds a global map of human
protein complexes. Mol. Syst. Biol.
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
