% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_report.R
\name{write_limma_plots}
\alias{write_limma_plots}
\title{Make interactive reports on differential abundance}
\usage{
write_limma_plots(
  DIAlist = NULL,
  grouping_column = NULL,
  output_dir = NULL,
  tmp_subdir = "tmp",
  key_column = NULL,
  height = 1000,
  width = 1000
)
}
\arguments{
\item{DIAlist}{A DIAlist object, with statistical results.}

\item{grouping_column}{The name of the column in the metadata which
gives information on how to group samples for the interactive
abundance plot.}

\item{output_dir}{The directory in which to create the reports and save the
plot files. No defaults, must be specified.}

\item{tmp_subdir}{The subdirectory within the output directory in which to
store temporary files. Deleted by default. Default is "tmp".}

\item{key_column}{Optional. The name of a column in the annotation data frame which gives
unique protein IDs. If provided, values form this column will appear as the title for the
protein intensity plots in the report. The column must contain unique values for each protein.
For example: key_column="Uniprot_ID" could be possible list of unique ids in
your annotations.}

\item{height}{The height of the interactive report objects, in pixels.
Default is 1000.}

\item{width}{The width of the interactive report objects, in pixels.
Default is 1000.}
}
\value{
Invisibly returns the input DIAlist.
}
\description{
Creates and saves interactive HTML reports summarizing
differential abundance analyses for each contrast in the results slot of the DIAlist.
Creates one HTML report for each contrast. Also creates a subfolder containing static .pdf
versions of all interactive plots. Currently, overwrites any previous reports
or other files with the same name.
}
\examples{
\dontrun{
  # Using defaults
  write_limma_plots(DIAlist,
                    grouping_column = "treatment")

  # Adjust size of report
  write_limma_plots(DIAlist,
                    grouping_column = "treatment",
                    height = 1500,
                    width = 1500)

 # Customize output directory
 write_limma_plots(DIAlist,
                   grouping_column = "treatment",
                   output_dir = "DA_results")

 # Add titles to the intensity plot,
 # using the values in the protein_id column
 # in the annotation of the DIAlist
 write_limma_plots(DIAlist,
                   grouping_column = "treatment",
                   key_column = "protein_id")
}

}
