% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_report.R
\name{write_proteinorm_report}
\alias{write_proteinorm_report}
\title{Create proteinorm report}
\usage{
write_proteinorm_report(
  DIAlist,
  grouping_column = NULL,
  out_dir = NULL,
  file = NULL,
  overwrite = FALSE,
  suppress_zoom_legend = FALSE
)
}
\arguments{
\item{DIAlist}{A DIAlist.}

\item{grouping_column}{The name of the column in the metadata which
gives information on how to group samples for normalization. Must be supplied:
some metrics can't be calculated for only one group.}

\item{out_dir}{The directory in which to save the report. If not provided,
will default to "protein_analysis/01_quality_control".}

\item{file}{The file name of the report to be saved. Must end in .pdf. Will
default to "proteiNorm_Report.pdf" if no filename is provided.}

\item{overwrite}{Should report file be overwritten if it already exists?
Default is FALSE.}

\item{suppress_zoom_legend}{Should the legend be removed from the zoomed
log2ratio plot? Default is FALSE}
}
\value{
Invisibly, the filename of the created report
}
\description{
Creates and saves as a PDF report a variety of plots which give
information about the performance of different normalization metrics.
}
\examples{
# No examples yet

}
\seealso{
\code{\link{norm_metrics}},
\code{\link{eval_pn_metric_for_plot}},
\code{\link{pn_plots_generic}},
\code{\link{pn_plots}},
}
