% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_report.R
\name{write_qc_report}
\alias{write_qc_report}
\title{Create quality control report}
\usage{
write_qc_report(
  processed_data,
  chosen_norm_method = NULL,
  grouping_column = NULL,
  label_column = NULL,
  enrich = c("protein", "phospho"),
  out_dir = NULL,
  file = NULL,
  overwrite = FALSE,
  top_proteins = 500,
  standardize = TRUE,
  pca_axes = c(1, 2),
  clust_metric = "euclidean",
  clust_method = "complete"
)
}
\arguments{
\item{processed_data}{The output of the \code{\link{process_data}} function:
a list object containing processed data and sample information..}

\item{grouping_column}{The name of column within the targets data frame which
gives information on how to group samples for normalization. If not supplied,
will warn the user and treat all samples as belonging to the same group.}

\item{enrich}{What type of analysis is this for? Options are
"protein" and "phospho". Only used for making output dir name when one isn't
supplied.}

\item{out_dir}{The directory in which to save the report. If not provided,
will default to either "protein_analysis/01_quality_control" or
"phospho_analysis/01_quality_control" within the current working
directory, depending on the enrich argument.}

\item{file}{The file name of the report to be saved. Must end in .pdf. Will
default to "QC_Report.pdf" if no filename is provided.}

\item{overwrite}{Should report file be overwritten if it already exists?
Default is FALSE.}

\item{norm.method}{The normalization method for which to perform the QC analysis.
Should be a name of one of the datasets in normList.}

\item{legend}{Include legends in the plots within the report? Default is TRUE.}

\item{top.proteins}{The number of most variable proteins to use for the analysis.
Default is 500.}

\item{stdize}{Should input data be standardized to a mean of 0 and std.dev of
1? If input data are not yet standardized, should be TRUE. Default is TRUE.}

\item{pca.axes}{A numeric vector of length 2 which lists the PC axes to plot.
Default is c(1,2), to plot the first two principal components.}

\item{pca.xlim}{Optional. Custom x-axis limits of the PCA plot. By default,
the min is 10\% below the min PC score on the x-axis, and the max is above
the max PC score on the x-axis, with some padding for sample labels.}

\item{pca.ylim}{Optional. Custom y-axis limits of the PCA plot. Default is
10\% above and below the min/max PC score on the y-axis.}

\item{pca.dot}{Character expansion factor for the size of the points in the
PCA plot. Default is 2.}

\item{clust.label.cex}{Character expansion factor for sample labels in the
dendrogram. Default is 1.}
}
\value{
Invisibly returns a list with three slots: \enumerate{
\item "plots"- A large list, where each element in the list is the returned
object from the corresponding plotting function for that type of plot.
\item "stats"- A dataframe with statistics on the data.
\item "param"- A dataframe giving the parameters used for making the report.
}
}
\description{
Creates, and optionally saves as a PDF report, a bunch of plots which give
information on the distribution, clustering, and correlation of protein intensities
across samples for the chosen normalization method. By default, the PCA and
clustering analyses are performed on the top 500 most variable proteins. This
function is a wrapper that calls many subfunctions: \itemize{
\item Makes boxplots of per-sample intensities
with \code{\link{qc_boxplot}}.
\item Makes violin plots of per-sample intensities
with \code{\link{qc_violin_plot}}.
\item Performs and plots a PCA with \code{\link{qc_pca_plot}}.
\item Does hierarchical clustering and plots a dendrogram with
\code{\link{qc_dendrogram}}.
\item Plots a correlation heatmap with \code{\link{qc_corr_hm}}.
} See the documentation of these subfunctions for more info.
}
\examples{
# No examples yet


}
