% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zscore_m.R
\name{Zscore_m}
\alias{Zscore_m}
\title{Zscore_m}
\usage{
Zscore_m(datInput, iter = 3)
}
\arguments{
\item{datInput}{Data frame with column names:
Experiment.id, Bait, and Prey.}

\item{iter}{Number of bootstraps to execute to estimate z scores.
Defaults to 3.}
}
\value{
Data frame containing bait-prey pairs with normalized Z-score.
}
\description{
This function applies Z-score to score instances
(e.g., bait-prey interactions (BPIs) in the data.frame. This scoring
algorithm is based on matrix model that treats bait–prey and
prey–prey interactions equally.
}
\examples{
data("Testdata_1")
Testdata_1 <- Testdata_1[, c(1,3:4)]
scoredfile1 <- Zscore_m(Testdata_1, iter = 2)
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
