% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_design.R
\name{add_design}
\alias{add_design}
\title{Prepare limma model design matrix}
\usage{
add_design(DAList, design_formula = NULL)
}
\arguments{
\item{DAList}{The DAList of normalized proteins.}

\item{design_formula}{A string for the design matrix using intercept, no intercept,
additive, or interaction models. See examples below.}
}
\value{
A DAList object with a design.
}
\description{
A design matrix is a model matrix of explanatory variables of a set of objects. Each row represents
individual samples and the columns represent the sample groups and factors. This function utilizes the
function \link[limma:modelMatrix]{limma::modelMatrix}.
}
\examples{
\dontrun{
# if x is a numerical covariate, then ~x and ~0+x are different models where the
# second model assumes the expected response is zero when x is zero.
# if A is a factor, then ~A and ~0+A are the same model, just parametrized slightly different.

# Intercept Model
data <- add_design(data, "~ group")

# No Intercept Model
data <- add_design(data, "~ 0 + group")

# Additive Model
data <- add_design(data, "~ 0 + group + batch")
data <- add_design(data, "~ 0 + group + pair_sample") # fixed effects
data <- add_design(data, "~ 0 + group + (1 | xyz)")   # mixed effects (fix + random effect)

# Interaction Model
data <- add_design(data, "~ group*gender")

}


}
