% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata.R
\name{add_metadata}
\alias{add_metadata}
\title{Add metadata to a DAList}
\usage{
add_metadata(DAList, metadata_file)
}
\arguments{
\item{DAList}{A DAList to which metadata will be added.}

\item{metadata_file}{A metadata file defining sample information necessary for the statistical design.}
}
\value{
An S3 list type with added metadata.
}
\description{
Add a data frame of metadata to a DAList object. The metadata file defines the
sample labels, groups, and other factors (such as batch, gender, age, paired samples, etc.)
required for the analysis and design matrix.
}
\examples{
\dontrun{

data <- add_metadata(data, "path/to/metadata_file.csv")

# An example metadata file contains columns: number, sample, batch, group. The number column
in the metadata must match the sample name in the data file. For instance, metadata file number "1"
corresponds to data column "Sample_01". This matching defines the other factors about the sample.
}

}
