% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_targets.R
\name{add_metadata}
\alias{add_metadata}
\title{Make a targets file}
\usage{
add_metadata(DIAlist, metadata_file)
}
\arguments{
\item{DIAlist}{A DIAlist to which metadata will be added}

\item{metadata_file}{A metadata file, giving info on the samples.}
}
\value{
A prototype of our new S3 list type, with added metadata
}
\description{
Creates a dataframe of "targets". Output depends on whether a metadata file
is supplied, and whether the metadata and targets can be matched successfully
(see Value, below). Calls two subfunctions: \code{\link{import_meta}} and
\code{\link{get_dia_sample_number}}.
}
\examples{
# No examples yet

}
