% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_methods.R
\name{apply_all_normalizations}
\alias{apply_all_normalizations}
\title{Apply all normalizations to a dataframe}
\usage{
apply_all_normalizations(data)
}
\arguments{
\item{data}{A dataframe of raw data to be normalized. Rows are proteins and
columns are raw intensity data.}
}
\value{
A list length 8, where each item in the list is a
named dataframe. Names give the normalization method, and the dataframe
gives the normalized intensity data
}
\description{
Takes in raw intensity data and applies 8 different normalization methods,
returning a list of the data normalized with each method. Used internally in
\code{\link{write_proteinorm_report}}.
}
\examples{
# No examples yet

}
\keyword{internal}
