% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_report.R
\name{apply_all_normalizations}
\alias{apply_all_normalizations}
\title{Apply all normalization methods to a set of raw data}
\usage{
apply_all_normalizations(data)
}
\arguments{
\item{data}{A dataframe or matrix of raw data to be normalized. Rows are proteins and
columns are raw intensity data.}
}
\value{
A list length 8, where each item in the list is a
named matrix. Names give the normalization method, and the matrix
contains the normalized data.
}
\description{
Takes in raw intensity data and applies 8 different normalization methods,
returning a list of the data normalized with each method. Used internally in
\code{\link{write_proteinorm_report}}.
}
\keyword{internal}
