% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_tables.R
\name{build_statlist}
\alias{build_statlist}
\title{Build a cleaned statlist for export to Excel (renaming z_scores)}
\usage{
build_statlist(
  DAList,
  stat_cols = c("logFC", "P.Value", "adj.P.Val", "sig.PVal", "sig.FDR", "movingSDs",
    "logFC_z_scores")
)
}
\arguments{
\item{stat_cols}{Character vector of column names to include from the results. Defaults include additional stats from tags.}

\item{results}{A named list of data frames. Each element is a contrast result (e.g., DAList$results$contrast_name).}

\item{movingSDs}{A named list of numeric vectors for each contrast (e.g., DAList$tags$movingSDs).}

\item{z_scores}{A named list of numeric vectors for each contrast (e.g., DAList$tags$logFC_z_scores).}
}
\value{
A named list of filtered data frames per contrast.
}
\description{
Build a cleaned statlist for export to Excel (renaming z_scores)
}
