% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_checks.R
\name{check_data}
\alias{check_data}
\title{check if x is a data.frame or matrix}
\usage{
check_data(x)
}
\arguments{
\item{x}{an object to be tested. (e.g. counts, metadata, annotation)}
}
\value{
The function returns TRUE if x is a data.frame or matrix
If x is not a data.frame or matrix the function returns FALSE.
}
\description{
for internal use. helper function. NEED TO UPDATE
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}

}
