% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_checks.R
\name{check_file}
\alias{check_file}
\title{Check if a file exists}
\usage{
check_file(x)
}
\arguments{
\item{x}{A character string representing a file path.}
}
\value{
Logical. Returns \code{TRUE} if \code{x} is a valid file path and the file exists.
Returns \code{FALSE} if \code{x} is not a valid string or the file does not exist.
}
\description{
Internal helper function to verify if an input string corresponds to an existing file.
}
\examples{
\dontrun{
if (interactive()) {
  check_file("path/to/existing_file.txt") # TRUE if file exists
  check_file("nonexistent_file.txt")      # FALSE
}
}
}
