% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_checks.R
\name{check_grouping_column}
\alias{check_grouping_column}
\title{metadata group column validator.}
\usage{
check_grouping_column(metadata, grouping_column, verbose = FALSE)
}
\arguments{
\item{metadata}{a data.frame of sample metadata. Rows are samples.}

\item{grouping_column}{The name of column or column index within
the metadata data.frame which gives information on how to group the samples.}

\item{verbose}{logical. Should the grouping_column be checked for
character length? Default: False}
}
\value{
If all checks pass the function returns the {.arg grouping_column}
column name.
}
\description{
Internal function for validating metadata grouping column information
}
\examples{

\dontrun{

metadata <- data.frame(sample = c("C1","C2","T1","T2"),
                       sample_id = c("C100","C202","T303","T100"),
                       condition = c("Control","Control","Treatment","Treatment"))

metadata2 <- data.frame(sample = c("C1","C2","T1","T2"),
                        sample = c("C100","C202","T303","T100"),
                        condition = c("Control","Control","Treatment","Treatment"),
                        check.names = FALSE)

metadata3 <- data.frame(sample = c("C2","C2","T1","T2"),
                        sample_id = c("C100","C202",NA,"T100"),
                        sample_name = c("sample_1","sample_2","","sample_4"),
                        condition = c("Baseline_Saline_Control",
                                      "Baseline_Saline_Control",
                                      "Treatment","Treatment"),
                        treatment = c("","Control","Chemo","Chemo"),
                        group = c("Control",NA,"Treatment",""),
                        chemo = c("Control","Saline","Treatment","Treatment"),
                        check.names = FALSE)


check_grouping_column(metadata = metadata,  grouping_column = "condition")
check_grouping_column(metadata = metadata,  grouping_column = "replicates")
check_grouping_column(metadata = metadata,  grouping_column = 7)
check_grouping_column(metadata = metadata2, grouping_column = 3)
check_grouping_column(metadata = metadata2, grouping_column = "sample")
check_grouping_column(metadata = metadata3, grouping_column = "condition")
check_grouping_column(metadata = metadata3, grouping_column = 5)
check_grouping_column(metadata = metadata3, grouping_column = "group")
check_grouping_column(metadata = metadata4, grouping_column = 7)

}
}
\keyword{internal}
