% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_checks.R
\name{check_int}
\alias{check_int}
\title{function identifies positive integer values}
\usage{
check_int(x)
}
\arguments{
\item{x}{description x a value to be tested. x must be single value of
length = 1L The function identifies positive whole numbers including 0
https://www3.ntu.edu.sg/home/ehchua/programming/howto/Regexe.html}
}
\description{
this function does some stuff NEED TO UPDATW
}
\examples{

\dontrun{
if(interactive()){

df <- data.frame(c1 = c(1, "ABC", Inf),
                 c2 = c(NA, 4.3, 6))

sapply(unlist(df), check_int)

mat <- matrix(c(1, NA, 2, Inf, -2, 10),
              nrow = 3, ncol = 2, byrow = TRUE,
              dimnames = list(c(paste0("gene_",1:3)), c("C1","C2")))

sapply(c(mat), check_int)

c(mat)[sapply(c(mat), check_int) == FALSE]

}
}


}
