% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_checks.R
\name{check_logical}
\alias{check_logical}
\title{checks that x is a TRUE or FALSE logical of length 1.}
\usage{
check_logical(x)
}
\arguments{
\item{x}{a value to be checked.}
}
\value{
the function returns TRUE if x is a TRUE OR FALSE logical of length 1.
the function returns FALSE if x is any of the following numeric,
has length > 1L, is an empty string, a series of blank spaces,
NA, NULL, or a character string.
}
\description{
for internal use. helper function. NEED TO UPDATE
}
\examples{

\dontrun{
if(interactive()){

check_logical(x = "ProjectID_122522")
check_logical(x = 123)
check_logical(x = 1.2342)
check_logical(x = NA)
check_logical(x = NULL)
check_logical(x = c("oh","snap"))
check_logical(x = TRUE)
check_logical(x = FALSE)
check_logical(x = c(TRUE,TRUE))
check_logical(x = "TRUE")
check_logical(x = "     ")
check_logical(x = "")
check_logical(x = Inf)

}
}


}
