% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_checks.R
\name{check_long}
\alias{check_long}
\title{Identify Values Above a Specified Character Length}
\usage{
check_long(x, char_thresh = 15, na.rm = FALSE)
}
\arguments{
\item{x}{A vector of values. This parameter is typically a vector of column names,
row names, or values in a particular metadata column.}

\item{char_thresh}{A numeric value (greater than or equal to 0). This defines the
maximum allowed character length for values in \code{x}. Any value exceeding this length
will be returned.}

\item{na.rm}{A logical value indicating whether to remove \code{NA} values from \code{x}. Default is \code{FALSE}.}
}
\value{
The function returns a vector of values in \code{x} that exceed the specified
\code{char_thresh} character length. If no such values exist, it returns \code{NULL}. If
\code{na.rm = FALSE}, \code{NA} and \code{NaN} values in \code{x} will be included in the output.
\subsection{x(no NAs) + na.rm = TRUE/FALSE + all below ==> NULL}{
}

\subsection{x(NAs) + na.rm = TRUE + all below ==> NULL}{
}

\subsection{x(NAs) + na.rm = FALSE + all below ==> c(NA, NaN)}{
}

\subsection{x(no NAs) + na.rm = TRUE/FALSE + above == > c(values)}{
}

\subsection{x(NAs) + na.rm = TRUE + above ==> c(values)}{
}

\subsection{x(NAs) + na.rm = FALSE + above ==> c(values, NA, NaN)}{
}
}
\description{
This function identifies values in a vector that exceed a specified maximum number
of characters. This is often used to identify long sample names or group labels that
could cause issues in visualizations or presentations.
}
\examples{
\dontrun{
if(interactive()){
  x <- c("Con_01", "Con_02", NA, "TAC_Treatment_01", "", Inf, NaN, "TAC_Treat_02")
  y <- c("Con_01", "Con_02", "Treat_01", "Treat_101")
  z <- c("Con_01", NA, "TAC_Treatment_01", "TAC_Treat_02", "", Inf)
 ## identify values with more than 15 characters
  check_long(x = x, char_thresh = 15, na.rm = FALSE)
 ## identify values with more than 8 characters
  check_long(x = y, char_thresh = 8)
 ## identify values with more than 6 characters
  check_long(x = y, char_thresh = 6)
}
}

}
