% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_checks.R
\name{check_metadata}
\alias{check_metadata}
\title{Validate Metadata Structure}
\usage{
check_metadata(x, verbose = FALSE)
}
\arguments{
\item{x}{A \code{data.frame} containing sample metadata. Each row represents a sample.}

\item{verbose}{A logical value indicating whether to check if row and column names exceed
a character length threshold. Default is \code{FALSE}.}
}
\value{
Invisibly returns the validated input \code{data.frame}.
}
\description{
This internal function checks the validity of a metadata \code{data.frame}, ensuring
it meets structural requirements such as unique column names, non-empty row names,
and appropriate formatting.
}
\details{
\itemize{
\item Ensures \code{x} is a \code{data.frame} and not empty.
\item Checks that column names are present, unique, and free from \code{NA} or blank values.
\item Ensures row names are not missing, blank, or duplicated.
\item If \code{verbose = TRUE}, warns if row or column names exceed 15 characters.
}
}
\examples{
\dontrun{
if(interactive()){

metadata <- data.frame(sample = c("C1", "C2", "T1", "T2"),
                       sample_id = c("C100", "C202", "T303", "T100"),
                       condition = c("Control", "Control", "Treatment",
                                     "Treatment"))

metadata2 <- data.frame(sample = c("C1", "C2", "T1", "T2"),
                        sample = c("C100", "C202", "T303", "T100"),
                        condition = c("Control", "Control", "Treatment",
                                      "Treatment"),
                        check.names = FALSE)

metadata3 <- data.frame(long_sample_column_name = c("C1", "C2", "T1", "T2"),
                        sample_id = c("C100", "C202", "T303", "T100"),
                        condition = c("Control", "Control", "Treatment",
                                      "Treatment"),
                        row.names = c("C100", "C202", "T303",
                                      "T100_long_row_name"))

metadata4 <- data.frame(sample_id = c("C1", "C2", "T1", "T2"),
                        sample = c("C100", "C202", "T303", "T100"),
                        condition = c("Control", "Control", "Treatment",
                                      "Treatment"),
                        row.names = c("C1", "C2", "T2", ""), check.rows = FALSE)


metadata  <- check_metadata(x = metadata, verbose = TRUE)
metadata2 <- check_metadata(x = metadata2, verbose = TRUE)
metadata3 <- check_metadata(x = metadata3, verbose = TRUE)
metadata4 <- check_metadata(x = metadata4, verbose = TRUE)

 }
}

}
\keyword{internal}
