% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_checks.R
\name{check_num}
\alias{check_num}
\title{function identifies positive numeric values}
\usage{
check_num(x)
}
\arguments{
\item{x}{a value to be tested. x must be value of length = 1L}
}
\value{
the function returns TRUE if x is a positive numeric value
greater than or equal to zero. The function returns FALSE if x is
any of the following: non-numeric, a character string, vector with
multiple values (length > 1L), an empty string (blank), white space
e.g. blank spaces, tabs, new lines), NA, NULL, Inf, or a number less
than zero i.e. negative number etc.
}
\description{
this function is used to check that the norm.factors are valid numeric values
}
\examples{

\dontrun{
if(interactive()){

check_num(x = "snap")
check_num(x = 1.4323)
check_num(x = c(12,130))
check_num(x = NA)
check_num(x = -45)
check_num(x = Inf)
check_num(x = 0)
check_num(x = "    ")
check_num(x = " \n")
check_num(x = "")
check_num(x = 1e6)
check_num(x = 1/12)

 }
}


}
