% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_checks.R
\name{check_string}
\alias{check_string}
\title{checks that x is a character string of length 1.}
\usage{
check_string(x)
}
\arguments{
\item{x}{a value to be checked.}
}
\value{
the function returns TRUE if x is a character string of length 1.
the function returns FALSE if x is any of the following numeric,
has length > 1L, is an empty string, series of blank spaces,
NA, NULL, Inf etc.
}
\description{
for internal use. helper function. NEED TO UPDATE
}
\examples{

\dontrun{
if(interactive()){

check_string(x = "ProjectID_122522")
check_string(x = 123)
check_string(x = NA)
check_string(x = NULL)
check_string(x = c("oh","snap"))
check_string(x = TRUE)
check_string(x = "     ")
check_string(x = "")
check_string(x = Inf)

 }
}


}
