% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_checks.R
\name{check_vec}
\alias{check_vec}
\title{Check if a vector x includes a set of reference values.}
\usage{
check_vec(x, ref)
}
\arguments{
\item{x}{a vector of values to be checked. The length of x
must be equal to or greater than the ref vector with
no duplicate values matching the reference. This parameter
is typically the values from a particular metadata
or annotation column.}

\item{ref}{a vector of allowed values. All values in the ref
set must be unique with no NA or empty values. This parameter
is typically the column names or row names of a counts matrix.}
}
\value{
The function returns TRUE if the check is successful
and FALSE if the check is unsuccessful.
}
\description{
General purpose function that checks if a reference vector is a
subset of a given input vector x.
}
\details{
x should have the same or longer length as the number of values
in ref with no duplicated values. The values in x and ref must
follow R's syntax rules for variable names (i.e. column names
and row names).

See \code{\link{base::make.names}} for naming requirements.
}
\examples{

\dontrun{
if(interactive()){

## ref is a subset of x
x   <- paste0("gene_", 1:10)
ref <- paste0("gene_", 3:7)
check_vec(x = x, ref = ref)

## x does not include all ref values
x   <- paste0("gene_", 1:5)
ref <- paste0("gene_", 4:6)
check_vec(x = x, ref = ref)

## x does not include any ref values
x   <- paste0("gene_", 1:5)
ref <- paste0("gene_", 20:22)
check_vec(x = x, ref = ref)

## x is < ref
## x contains duplicates
## x contains NA
## ref includes duplicates
## ref includes NA
## ref includes empty strings
## ref includes blank spaces

## ref contains values equal to blank spaces


samples <- data.frame(sample_id = c("C100","C202","T303","T100"),
                      sample    = c("C1","C2","T1","T2"),
                      group     = c("Con","Con","Treat","Treat"))

counts <- data.frame(C100 = c(0,10,21,3,4),
                     C202 = c(0,0,0,2,1),
                     T100 = c(15,33,21,55,42),
                     row.names = c(paste0("gene_",1:5)))

genes <- data.frame(gene_id   = c(paste0("gene_",1:5),"NA.","NA..1",""),
                    gene_name = LETTERS[1:8])


is_id <- unlist(lapply(samples,
                       function(x) {
                         check_vec(x = x, ref = colnames(counts))
                     }))

is_id <- sapply(as.list.data.frame(samples),
                function(x) {
                  check_vec(x = x, ref = colnames(counts))
                })

id_column <- colnames(samples)[is_id][1]

sort(samples[, id_column]); sort(colnames(counts))


## find id column in annotation matching counts row names
is_id <- unlist(lapply(genes,
                       function(x) {
                         check_vec(x = x, ref = row.names(counts))
                     }))


 }
}
}
