% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_tables.R
\name{create_combined_results}
\alias{create_combined_results}
\title{Write combined results csv}
\usage{
create_combined_results(annotation, data, statlist)
}
\arguments{
\item{annotation}{A dataframe of annotation data for each gene/protein.}

\item{data}{A dataframe of normalized intensity data for each sample.}

\item{statlist}{A list of per-contrast DE results.}
}
\value{
A dataframe of the combined results.
}
\description{
Internal function used to write the combined statistical results to a .csv
file.
}
\keyword{internal}
