% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class.R
\name{DAList}
\alias{DAList}
\title{Create a DAList}
\usage{
DAList(
  data,
  annotation,
  metadata,
  design = NULL,
  eBayes_fit = NULL,
  results = NULL,
  tags = NULL
)
}
\arguments{
\item{data}{A dataframe or matrix containing protein intensity data for each sample. Rows are proteins, columns are samples.}

\item{annotation}{A dataframe containing protein annotation data. REQUIRED COLUMNS??}

\item{metadata}{A dataframe containing metadata on each sample. REQUIRED COLUMNS?}

\item{design}{Optional, possibly unneeded. But a list of design information.}

\item{eBayes_fit}{Optional, possibly unneeded. An ebayes fit/MAarray object from limma ebayes}

\item{results}{Optional, possibly unneeded. list of table of stats.}

\item{tags}{Optional, possibly unneeded. List of tags.}
}
\value{
A DAList object
}
\description{
A function to create a DAList from existing data. Need to add in notes and further
documentation as we decide on exactly what requirements we have.
In particular, should note that: data and annotation need to have the same number of
rows and should already be in order. Columns in data should be same as number of rows in metadata.
TODO: need to expand on documentation as we finalize things. Also need to decide if we
should include all the other slots (design through tags). Also need to add examples.
}
