% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_plotting.R
\name{eval_pn_metric_for_plot}
\alias{eval_pn_metric_for_plot}
\title{Evaluate a normalization metric across a list of normalized data}
\usage{
eval_pn_metric_for_plot(
  normList,
  grouping,
  metric = c("PCV", "PMAD", "PEV", "COR", "log2ratio")
)
}
\arguments{
\item{normList}{A named list of normalized data matrices.}

\item{grouping}{A character or factor vector, listing the group(s) the samples
belong to.}

\item{metric}{The normalization metric to calculate. Can be "PCV",
"PMAD", "PEV", or "COR". See \code{\link{norm_metrics}}.}
}
\value{
A data frame containing the selected metric,
ready to be used for plotting.
}
\description{
Applies a metric (see \code{\link{norm_metrics}}) across a list of normalized
data matrices (most likely produced by \code{\link{apply_all_normalizations}})
and outputs the results in a data frame ready to be used with internal plotting
functions (see \code{\link{pn_plots_generic}}).
}
\keyword{internal}
