% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{extract_protein_data}
\alias{extract_protein_data}
\title{Extract protein data}
\usage{
extract_protein_data(annotation_data)
}
\arguments{
\item{annotation_data}{Annotation data from which to extract protein data.}
}
\value{
A  data frame of cleaned up annotation data, with protein info
}
\description{
Third subfunction called by \code{\link{read_DIA_data}}. Processes the raw
protein annotation strings to extract standard gene names,
accession numbers, and Ids.
}
\examples{
# No examples yet
}
