% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{extract_protein_data}
\alias{extract_protein_data}
\title{Extract protein data}
\usage{
extract_protein_data(data, sample_IDs = NULL)
}
\arguments{
\item{data}{Data from which to extract protein data. In the
\code{\link{read_DIA_data}} function, this is \code{qfilterData$data}, which
is the "data" slot of the result of the \code{\link{remove_contaminants}}
output: the data frame of both protein and intensity data.}

\item{sample_IDs}{Optional: a vector of sample IDs to analyze. Default is NULL, in which
case all sample IDs are analyzed. These are determined from column names.}
}
\value{
A list with 9 slots:
\enumerate{
\item "rawData": A dataframe, where each row is a proetin and each column
is a sample. Not actually raw, I think(?): at this point in the pipeline,
contaminants have been removed. Rownames in this data frame give a protein ID,
in the form of UniprotID_Gene_name_id.
\item "rawAnnot"-A dataframe, where each row is a protein and the columns
are further protein annotation information. Rownames in this data frame give
a protein ID, in the form of UniprotID_Gene_name_id.
\item "sampleIDs"- A vector of the sample IDs for the data. Seems like the
column names from the rawData slot. Redundant?
\item "annotColumns"- The column names of the "rawAnnot" slot. Redundant?
\item "num_samples"- The number of samples.
Currently seems redundant: just ncol() of rawData slot. Maybe useful if we
moved back to merging annotation and sample data?
\item "num_extracted"- The number of proteins (nrow() of rawData slot).
Redundant?
\item "pipe"- The pipeline used, "DIA".
\item "enrich"- The enrichment type, "protein".
\item "pattern"- Unclear what this is at the moment. Comes from the
results of the \code{\link{extract_sampleIDs}} function. For DIA data, though,
there is no "pattern" slot in the result list from
\code{\link{extract_sampleIDs}}. For TMT and other data, seems to be the the
regex pattern that was used for getting sample IDs from columns?
}
}
\description{
Third subfunction called by \code{\link{read_DIA_data}}. Processes the raw
protein annotation strings to extract standard gene names,
accession numbers, and Ids.
}
\examples{
# No examples yet
}
