% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{extract_sampleIDs}
\alias{extract_sampleIDs}
\title{Extract sample IDs}
\usage{
extract_sampleIDs(colNames, sample_IDs = NULL)
}
\arguments{
\item{colNames}{A list of column names.}

\item{sample_IDs}{Optional: a vector of sample IDs to analyze. Default is NULL, in which
case all sample IDs are analyzed. These are determined from column names.}
}
\value{
A list, with a single slot. I think a vector of sample IDs.
}
\description{
A subfunction called by \code{\link{extract_protein_data}}. Does two
different things. In the "default" case, with sample_IDs = NULL, it takes in a
vector of column names (in our pipeline, these are the column names from
\code{qfilterData$data}, which is the "data" slot of the result
of the \code{\link{remove_contaminants}} output and is a data frame
of both protein and intensity data). Then, it extracts sample names from those
column names. If a vector of sample_IDs are supplied, it checks to see if those
IDs are present in the column names. If so, it returns them. If not, gives an
error.
}
\examples{
# No examples yet
}
