% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{filter_data}
\alias{filter_data}
\title{Filter protein data.}
\usage{
filter_data(
  data,
  targets,
  min.reps = NULL,
  min.grps = NULL,
  group_col = "group"
)
}
\arguments{
\item{data}{Raw data to be processed. In the pipeline, this is generally the
"data" slot of the list output by \code{\link{read_DIA_data}}.}

\item{targets}{An inputs target dataframe. In the pipeline, this is generally
the subsetted targets data frame (that is, with pools or other non-desired
samples removed).}

\item{min.reps}{For filtering, the minimum number of samples within a group
that need to have non-zero intensity for a given protein/peptide for that
peptide to be considered as quantified within a group.}

\item{min.grps}{For filtering, the minimum number of groups that must have at
least min.reps non-zero samples for a given protein/peptide to be retained.}

\item{group_col}{Optional. The name of the column in the targets dataframe which
lists the experimental group names. Default is "group"}
}
\value{
A list with 6 elements:
\enumerate{
\item "data"- A dataframe of the samples and proteins that were retained
for further analysis
\item "targets" - A dataframe of targets. Should be nearly the same as
the targets dataframe that was input, possibly with new rownames.
\item "nonzeroSamplesPerGroup"- A data frame listing the number of samples
with nonzero intensities for each protein (row) in each group (column).
\item "rm.data"- A dataframe of the proteins/rows filtered out.
\item "param"- A dataframe giving the parameters and arguments used for
filtering
\item "stats"- A dataframe giving statistics on filtering.
}
}
\description{
A subfunction of \code{\link{process_data}}, this function removes samples that
are not present in the targets file and filters out proteins/rows that aren't
present in enough samples. For filtering, data rows (proteins/peptides) are
only retained if they have non-zero intensities in at least min.reps samples
and in at least min.grps groups. E.g., with min.reps = 2 and min.groups = 1,
a protein must be quantified in two samples within one group to be retained.
This function calls one subfunction, \code{\link{make_log}}.
}
\examples{
# No examples yet


}
