% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_proteins.R
\name{filter_proteins_by_annotation}
\alias{filter_proteins_by_annotation}
\title{Remove proteins by annotations}
\usage{
filter_proteins_by_annotation(DIAlist, condition)
}
\arguments{
\item{DIAlist}{A DIAlist object to be filtered}

\item{condition}{A logical expression indicating which samples to keep.}
}
\value{
A DIAlist with some proteins filtered out
}
\description{
NEED TO REWRITE
}
\examples{
# No examples yet

}
