% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_proteins.R
\name{filter_proteins_contaminants}
\alias{filter_proteins_contaminants}
\title{Remove protein contaminant}
\usage{
filter_proteins_contaminants(DAList)
}
\arguments{
\item{DAList}{A DAList object.}
}
\value{
A DAList, with contaminant proteins removed.
}
\description{
THIS FUNCTION FOR UAMS INTERNAL USE. This function removes protein contaminants
based on aspects of protein names that are specific to UAMS-formatted data.
Specifically, removes proteins which contain the strings "DECOY" or "Group of"
in the Protein.Name column, derived from the data output by Scaffold DIA.
}
\examples{
\dontrun{
filtered <- filter_proteins_contaminants(DAList)
}

}
\keyword{UAMS}
