% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_samples.R
\name{filter_samples}
\alias{filter_samples}
\title{Filter samples from a DIAlist}
\usage{
filter_samples(DIAlist, condition)
}
\arguments{
\item{DIAlist}{The DIAlist to be filtered}

\item{condition}{A logical expression indicating which samples to keep.}
}
\value{
A prototype of our new S3 list type.
}
\description{
NEED TO REWRITE
}
\examples{
\dontrun{
DIAlist <- filter_samples(DIAlist, group != "Pool")
}

}
