% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_targets.R
\name{filter_samples}
\alias{filter_samples}
\title{Filter samples from a DIAlist}
\usage{
filter_samples(DIAlist, filter_list, ignore.case = TRUE)
}
\arguments{
\item{DIAlist}{The DIAlist to be filtered}

\item{filter_list}{A named list describing how to filter targets dataframe.
The name for each element of the list gives a column to be filters, and
each element of the list should be a character vector of strings to search for
in the given column. See examples.}

\item{ignore.case}{Should the case of the strings in rm.vals be ignored? Is
TRUE by default, such that case does not have to match. If FALSE, case must
match for rows to be removed.}
}
\value{
A prototype of our new S3 list type.
}
\description{
Removes unwanted samples (rows) from the targets data frame This function is
basically \code{\link[base]{subset}} with extra functionality. It removes
rows if they contain certain values in a given column. By default, ignores
the case of strings in rm.val. Calls \code{\link{make_log}} as a subfunction.
}
\examples{
\dontrun{
DIAlist <- filter_samples(DIAlist
                          filter_list = list(group = "pool"
                                             sample = c("sampleA", "sampleB")))
}

}
