% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colors.R
\name{get_colors}
\alias{get_colors}
\title{Generate colors for Sample Groups}
\usage{
get_colors(group)
}
\arguments{
\item{group}{A vector representing the sample group classifications.
If not a factor, it is converted into one.}
}
\value{
A named vector of colors, where names correspond to unique group labels.
}
\description{
Assigns colors to each unique sample group in metadata. If the number of groups is 12 or fewer,
colors are assigned using \code{\link{colorGroup}}. For more than 12 groups, the function
uses the \pkg{Polychrome} to generate a distinct color palette.
}
\details{
\itemize{
\item If there are 12 or fewer unique groups, the function assigns colors using \code{\link{colorGroup}}.
\item If there are more than 12 groups, it leverages \pkg{Polychrome} to generate distinct colors. The function
uses \code{\link{binfcolors}} as seed colors to create a new color palette
instead of creating rainbow color palette using colorGroup() function
\item If \pkg{grDevices} is not installed, an error is thrown when handling more than 12 groups.
\item group must be a vector
}
}
\examples{
\dontrun {
# example code
group <- factor(c("A","B","C"))
get_colors(group) 

# example2 
group_color <- get_colors(group = results$metadata$group)

}

}
