% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_targets.R
\name{get_dia_sample_number}
\alias{get_dia_sample_number}
\title{Get DIA sample number from sample IDs}
\usage{
get_dia_sample_number(sample_IDs)
}
\arguments{
\item{sample_IDs}{A character vector of sample IDs, from which the targets
file is constructed. Generally, these are the column names of the
data slot in the object returned by \code{\link{read_DIA_data}}, e.g.,
\code{colnames(ext$data)}. Currently, these sample_IDs have a required format:
"Pool" samples must include the regex pattern \code{[P/p]ool_[[:digit:]]} in
the column name, and samples must contain the regex pattern
\code{[S/s]ample_[[:digit:]]} in the column name.}
}
\value{
A numeric vector of the extracted sample numbers
}
\description{
Extracts the sample number from a vector of sample IDs
}
\examples{
# No examples yet
}
