% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Not in operator}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{A vector of values to be matched.}

\item{table}{A vector of values to be matched against.}
}
\value{
A logical vector indicating if elements of \code{x} are not in \code{table}.
}
\description{
A convenience operator equivalent to \code{!(x \%in\% table)}., i.e., checks
whether elements of \code{x} are not in \code{table}.
}
