% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_targets.R
\name{import_meta}
\alias{import_meta}
\title{Import metadata file}
\usage{
import_meta(file, sampleIDs, pipe = c("DIA", "TMT", "phosphoTMT"))
}
\arguments{
\item{file}{The filename of the metadata file to import}

\item{sampleIDs}{A character vector of sample IDs, from which the targets
file is constructed. In the DIA pipeline, these are the column names of the
data slot in the object returned by \code{\link{read_DIA_data}}, e.g.,
\code{colnames(ext$data)}. Currently, these sampleIDs have a required format:
"Pool" samples must include the regex pattern \code{[P/p]ool_[[:digit:]]} in
the column name, and samples must contain the regex pattern
\code{[S/s]ample_[[:digit:]]} in the column name.}

\item{pipe}{Which analysis pipeline are you running? Options are "DIA", "TMT",
"phosphoTMT". Output varies slightly based on the pipeline you
choose.}
}
\value{
A dataframe of the imported metadata.
}
\description{
Imports a delimted metadata file, checking for some required columns.
}
\examples{
# No examples yet

}
