% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_targets.R
\name{import_meta}
\alias{import_meta}
\title{Import metadata file}
\usage{
import_meta(input_file)
}
\arguments{
\item{input_file}{A metadata file, giving info on the samples.}
}
\value{
A dataframe of the imported metadata.
}
\description{
Imports a delimited metadata file, checking for some required columns.
}
\examples{
# No examples yet

}
