% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_design.R
\name{make_design}
\alias{make_design}
\title{Prepare limma model design matrix}
\usage{
make_design(
  targets,
  group_column = NULL,
  factor_columns = NULL,
  paired_column = NULL,
  designformula = NULL
)
}
\arguments{
\item{targets}{A targets dataframe. In the pipeline, usually the "targets" slot
of the list output by \code{\link{process_data}}.}

\item{group_column}{The name of the column in the targets dataframe that
describes the main groups to be compared in the limma model.}

\item{factor_columns}{Optional. The name of the column(s) in the targets dataframe
that describes additional statistical factors for analysis. Default is NULL.}

\item{paired_column}{Optional. The name of the column in the targets dataframe
that lists paired samples, for use in limma's paired mixed effect model.}

\item{designformula}{Optional. A chracter string that can be coercted into a terms
object that is appropriate for input as the object for stats:model.matrix(). When
designformula is specified group_columns, factor_columns and paired_column is ignored.}
}
\value{
A list with three slots \enumerate{
\item "design"- The model design matrix for the specified factors.
\item "targets"- An updated targets data frame including only the grouping,
factor, and paired columns specified.
\item "designformula"- The model design formula for the specified factors.
}
}
\description{
Creates the model design matrix, targets data frame, and design formula for use in
limma. Either group_column or designformula must be specified. If designformula is
specific all other column parameters are ignored.
}
\examples{
# No examples yet

}
