% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_data_import.R
\name{make_log}
\alias{make_log}
\title{Make a log file}
\usage{
make_log(param, stats, title = "", save = TRUE)
}
\arguments{
\item{param}{The param object within the \code{\link{extract_data}} function
body. Slightly unclear what the requirements for this are.}

\item{stats}{The stats object within the \code{\link{extract_data}} function
body. Slightly unclear what the requirements for this are.}

\item{title}{A title for the printed parameter block. Default is "".}

\item{save}{TRUE/FALSE: should the parameters be save dot a text file?}
}
\value{
A list, with the param and stats objects. If SAVE = TRUE, has side
effect of creating log files.
}
\description{
Fourth subfunction called by \code{\link{extract_data}}. Used to write out
some logs during the data extraction process. Might want to rethink how this works.
}
\examples{
# No examples yet
}
