% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_targets.R
\name{make_targets}
\alias{make_targets}
\title{Make a targets file}
\usage{
make_targets(
  file = NULL,
  sampleIDs,
  pipe = c("DIA", "TMT", "phosphoTMT"),
  enrich = c("protein", "phospho")
)
}
\arguments{
\item{file}{Optional: a metadata file, giving info on the samples. Required
format and columns for metadata file vary by pipeline.
Need to document somewhere.}

\item{sampleIDs}{A character vector of sample IDs, from which the targets
file is constructed. In the DIA pipeline, these are the column names of the
data slot in the object returned by \code{\link{extract_data}}, e.g.,
\code{colnames(ext$data)}. Currently, these sampleIDs have a required format:
"Pool" samples must include the regex pattern \code{[P/p]ool_[[:digit:]]} in
the column name, and samples must contain the regex pattern
\code{[S/s]ample_[[:digit:]]} in the column name.}

\item{pipe}{Which analysis pipeline are you running? Options are "DIA", "TMT",
"phosphoTMT". Output varies slightly based on the pipeline you
choose.}

\item{enrich}{Another aspect of pipeline? Options are "protein" and "phospho".
Seems semi-redundant?: DIA, TMT, and LF seem to always be "protein" in the
code. PhosphoTMT can be either protein or phospho, though that seems to vary
across subfunctions a little?}
}
\value{
It depends:
\itemize{
\item If a metadata file is supplied and could be successfully matched
against the input column names, a full targets dataframe is returned,
and a targets.csv file is created in the working directory.
\item If a metadata file is supplied but matching against column names
was unsuccessful, returns, with a warning, a list with two slots:
(1) the dataframe of assembled targets, and (2) the dataframe of
imported metadata. These must be manually combined before further
analysis. No targets.csv file is created.
\item When a metadata file is not supplied, returns, with a warning, an
incomplete targets dataframe, and does not create a targets.csv file.
}
}
\description{
Creates a dataframe of "targets". Output depends on whether a metadata file
is supplied, and whether the metadata and targets can be matched successfully
(see Value, below). Calls two subfunctions: \code{\link{import_meta}} and
\code{\link{get_dia_sample_number}}.
}
\examples{
# No examples yet
}
