% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_report.R
\name{missingValueHeatmap}
\alias{missingValueHeatmap}
\title{Create a missing values heatmap}
\usage{
missingValueHeatmap(
  missing,
  groups,
  batch = NULL,
  column_sort = c("cluster", "group", "batch"),
  groupColors,
  batchColors,
  sampleLabels = NULL,
  legend = FALSE
)
}
\arguments{
\item{missing}{An input matrix describing the missing data. See internals
of \code{\link{plotHeatmapsForReport}}.}

\item{groups}{Optional, a vector describing what experimental or treatment
group each sample belongs to. Generally, the "group" column in the targets
data frame output by \code{\link{process_data}}. If not supplied, will
warn the user and treat all samples as belonging to the same group.}

\item{batch}{Optional, a vector describing what batch each sample belongs to.
If not supplied, will treat all samples as being from the same batch.}

\item{column_sort}{How should the columns of the heatmap be sorted? Options
are: "cluster"- sort by similarity in missing values, "group"- sort samples
by group, "batch"- sort samples by batch.}

\item{groupColors}{A vector of colors to use for coloring samples by group}

\item{batchColors}{A vector of colors to use for coloring samples by batch.}

\item{sampleLabels}{Optional, a set of sample labels to use. If not supplied,
defaults to using the column names in the normList.}

\item{legend}{Include a legend in the plot? Default is FALSE.}
}
\value{
A ComplexHeatmap object
}
\description{
Makes a ComplexHeatmap object showing a heatmap of missing values in the
input data.
}
\examples{
# No examples yet.

}
