% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_functions}
\alias{norm_functions}
\alias{logNorm}
\alias{medianNorm}
\alias{meanNorm}
\alias{vsnNorm}
\alias{quantNorm}
\alias{cycLoessNorm}
\alias{rlrNorm}
\alias{giNorm}
\title{Normalization functions}
\usage{
logNorm(dat)

medianNorm(logDat)

meanNorm(logDat)

vsnNorm(dat)

quantNorm(logDat)

cycLoessNorm(logDat)

rlrNorm(logDat)

giNorm(logDat)
}
\arguments{
\item{dat}{A numeric data frame, where each row is a protein and columns
are the raw, unstandardized intensities}

\item{logDat}{A numeric data frame, where each row is a protein and columns
are log2-transformed intensities.}
}
\value{
A matrix of normalized sample intensities
}
\description{
A set of functions that normalize sample data (in columns of a data frame
or matrix), according to various methods: \itemize{
\item logNorm- A binary (base2) log transformation
\item medianNorm- Divides by per-sample median, then multiplies by the
average of the per-sample medians.
\item meanNorm- Divides by per-sample mean, then multiplies by the
average of the per-sample means
\item vsnNorm-  Variance-stabilizing normalization (vsn),
using the \code{\link[vsn:justvsn]{vsn::justvsn}} function.
\item quantNorm-  Quantile normalization, using the
\code{\link[preprocessCore:normalize.quantiles]{preprocessCore::normalize.quantiles}}
function.
\item cycLoessNorm- Cyclic Loess normalization, using the
\code{\link[limma:normalizeCyclicLoess]{limma::normalizeCyclicLoess}}
function.
\item rlrNorm- Global linear regression normalization, using the
\code{\link[NormalyzerDE:performGlobalRLRNormalization]{NormalyzerDE::performGlobalRLRNormalization}}
function.
\item giNorm- Global intensity normalization, using the
\code{\link[NormalyzerDE:globalIntensityNormalization]{NormalyzerDE::globalIntensityNormalization}}
function.
}
}
\examples{
# No examples yet

}
