% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_methods.R
\name{norm_functions}
\alias{norm_functions}
\alias{log2Norm}
\alias{medianNorm}
\alias{meanNorm}
\alias{vsnNorm}
\alias{quantileNorm}
\alias{cycloessNorm}
\alias{rlrNorm}
\alias{giNorm}
\title{Normalization functions}
\usage{
log2Norm(dat)

medianNorm(logDat)

meanNorm(logDat)

vsnNorm(dat)

quantileNorm(logDat)

cycloessNorm(logDat)

rlrNorm(logDat)

giNorm(dat)
}
\arguments{
\item{dat}{A numeric data frame, where each row is a protein and columns
are the raw, unstandardized intensities}

\item{logDat}{A numeric data frame, where each row is a protein and columns
are log2-transformed intensities.}
}
\value{
A matrix of normalized sample intensities
}
\description{
A set of functions that normalize sample data (in columns of a data frame
or matrix), according to various methods: \itemize{
\item log2Norm- A binary (base2) log transformation
\item medianNorm- Divides by per-sample median, then multiplies by the
average of the per-sample medians.
\item meanNorm- Divides by per-sample mean, then multiplies by the
average of the per-sample means
\item vsnNorm-  Variance-stabilizing normalization (vsn),
using the \code{\link[vsn:justvsn]{vsn::justvsn}} function.
\item quantileNorm-  Quantile normalization, using the
\code{\link[preprocessCore:normalize.quantiles]{preprocessCore::normalize.quantiles}}
function.
\item cycloessNorm- Cyclic Loess normalization, using the
\code{\link[limma:normalizeCyclicLoess]{limma::normalizeCyclicLoess}}
function.
\item rlrNorm- Global linear regression normalization, inspired by the
\code{NormalyzerDE::performGlobalRLRNormalization}
function, but with a slightly different implementation.
\item giNorm- Global intensity normalization, inspired by the
\code{NormalyzerDE::globalIntensityNormalization}
function, but with a slightly different implementation. In brief,
normalizes each sample/column by dividing by the total intensity for the
sample and multiplying by the median of per-sample intensities. Thus,
like mean or median normalization, but using sum instead of
}
}
\examples{
# No examples yet

}
\keyword{internal}
