% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_functions}
\alias{norm_functions}
\alias{log2Norm}
\alias{medianNorm}
\alias{meanNorm}
\alias{vsnNorm}
\alias{quantileNorm}
\alias{cycloessNorm}
\alias{rlrNorm}
\alias{giNorm}
\title{Normalization functions}
\usage{
log2Norm(dat)

medianNorm(logDat)

meanNorm(logDat)

vsnNorm(dat)

quantileNorm(logDat)

cycloessNorm(logDat)

rlrNorm(logDat)

giNorm(dat)
}
\arguments{
\item{dat}{A numeric data frame, where each row is a protein and columns
are the raw intensities}

\item{logDat}{A numeric data frame, where each row is a protein and columns
are log2-transformed intensities.}
}
\value{
A matrix of normalized sample intensities
}
\description{
A set of functions that normalize data (in columns of a data frame
or matrix), according to various methods: \itemize{
\item log2Norm- Binary (base2) log transformation.
\item medianNorm- Divides each sample by the per-sample median, then multiplies all samples
by the average of the per-sample medians.
\item meanNorm- Divides each sample by the per-sample mean, then multiplies all samples
by the average of the per-sample means.
\item vsnNorm- Variance-stabilizing normalization (vsn),
using the \code{\link[vsn:justvsn]{vsn::justvsn}} function.
\item quantileNorm- Quantile normalization, using the
\code{\link[preprocessCore:normalize.quantiles]{preprocessCore::normalize.quantiles}}
function.
\item cycloessNorm- Cyclic Loess normalization, using the
\code{\link[limma:normalizeCyclicLoess]{limma::normalizeCyclicLoess}}
function.
\item rlrNorm- Global linear regression normalization, inspired by the
\code{NormalyzerDE::performGlobalRLRNormalization}
function, but with a slightly different implementation.
\item giNorm- Global intensity normalization, inspired by the
\code{NormalyzerDE::globalIntensityNormalization}
function, but with a slightly different implementation. Divides each sample by the
total intensity for each sample, then multiplies all samples by the median of per-sample
intensities.
}
}
\keyword{internal}
