% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_methods.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize data in a DIAlist}
\usage{
normalize_data(
  DIAlist,
  norm_method = c("log2", "median", "mean", "vsn", "quantile", "cycloess", "rlr", "gi")
)
}
\arguments{
\item{DIAlist}{A DIAlist with raw data to be normalized.}

\item{norm_method}{A normalization method to use. Options are "log2", "median",
"mean", "vsn", "quantile", "cycloess", "rlr", and "gi".}
}
\value{
A DIAlist with normalized data.
}
\description{
Normalizes the raw data in a DIAlist.
}
\examples{
# No examples yet

}
