% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize data in a DAList}
\usage{
normalize_data(
  DAList,
  norm_method = c("log2", "median", "mean", "vsn", "quantile", "cycloess", "rlr", "gi")
)
}
\arguments{
\item{DAList}{A DAList containing non-normalized data.}

\item{norm_method}{A normalization method to use. Options are "log2", "median",
"mean", "vsn", "quantile", "cycloess", "rlr", and "gi". Default is "log2".}
}
\value{
A DAList with normalized data.
}
\description{
Normalizes data in a DAList, using one of the available normalization options (see Details).
}
\details{
Available normalization methods: \itemize{
\item log2- Binary (base2) log transformation.
\item median- Divides each sample by the per-sample median, then multiplies all samples
by the average of the per-sample medians.
\item mean- Divides each sample by the per-sample mean, then multiplies all samples
by the average of the per-sample means.
\item vsn- Variance-stabilizing normalization (vsn),
using the \code{\link[vsn:justvsn]{vsn::justvsn}} function.
\item quantile- Quantile normalization, using the
\code{\link[preprocessCore:normalize.quantiles]{preprocessCore::normalize.quantiles}}
function.
\item cycloess- Cyclic Loess normalization, using the
\code{\link[limma:normalizeCyclicLoess]{limma::normalizeCyclicLoess}}
function.
\item rlr- Global linear regression normalization, inspired by the
\code{NormalyzerDE::performGlobalRLRNormalization}
function, but with a slightly different implementation.
\item gi- Global intensity normalization, inspired by the
\code{NormalyzerDE::globalIntensityNormalization}
function, but with a slightly different implementation. Divides each sample by the
total intensity for each sample, then multiplies all samples by the median of per-sample
intensities.
}
}
\examples{
\dontrun{
norm_data <- normalize_data(DAList, "log2")
}

}
