% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_normalization.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize raw intensity data}
\usage{
normalize_data(data, targets)
}
\arguments{
\item{data}{A dataframe of raw data to be normalized. Rows are proteins and
columns are raw intensity data.}

\item{targets}{A targets dataframe listing the samples to be analyzed.}
}
\value{
A list with two elements:
\enumerate{
\item "normList"- a list of length 8, where each item in the list is a
named dataframe. Names give the normalization method, and the dataframe
gives the normalized intensity data
\item "targets"- The targets dataframe that was passed into the function.
}
}
\description{
Takes in raw intensity data and applies 8 different normalization methods. A
subfunction of \code{\link{process_data}}. See \code{\link{norm_functions}}
for info on all the normalization functions.
}
\examples{
# No examples yet

}
