% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_plotting.R
\name{pn_plots}
\alias{pn_plots}
\alias{pn_plot_PCV}
\alias{pn_plot_PMAD}
\alias{pn_plot_PEV}
\alias{pn_plot_COR}
\alias{pn_plot_log2ratio}
\title{Functions for plotting normalization metrics}
\usage{
pn_plot_PCV(normList, grouping)

pn_plot_PMAD(normList, grouping)

pn_plot_PEV(normList, grouping)

pn_plot_COR(normList, grouping)

pn_plot_log2ratio(normList, grouping, zoom = F, legend = T)
}
\arguments{
\item{normList}{A named list of normalized data matrices. Generally, the "normList"
slot of the list that is output by \code{\link{process_data}}.}

\item{grouping}{A character or factor vector, listing the group(s) the samples
belong to.}

\item{zoom}{Should the plot cover the full range of log2ratios, or zoom
in around 0? Default is FALSE.}

\item{legend}{Should the plot include the legend? Default is TRUE.}
}
\value{
A ggplot object of the plot.
}
\description{
A set of functions that take in normalized sample data and a grouping factor
and calculate some metric of of variability, error, etc., that we can use
to evaluate normalization methods. See \code{\link{norm_metrics}} for explanations
of each metric.
}
\examples{
# No examples yet

}
