% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting.R
\name{qc_boxplot}
\alias{qc_boxplot}
\title{Boxplot for QC report}
\usage{
qc_boxplot(
  data,
  groups = NULL,
  sampleLabels = NULL,
  title = NULL,
  legend = TRUE
)
}
\arguments{
\item{data}{A data frame of intensity data, likely normalized. Rows should be
proteins and columns should be samples.}

\item{groups}{A character or factor vector, listing the group(s) the samples
belong to.}

\item{sampleLabels}{Optional, a vector of sample labels to use. If not supplied,
defaults to using the column names in the data.}

\item{title}{Optional, a plot title.}

\item{legend}{Include a legend in the plot? Default is TRUE.}
}
\value{
Invisibly returns the list object created by
\code{\link[graphics:boxplot]{graphics::boxplot}}.
}
\description{
Makes a boxplot of per-sample intensities. Samples are grouped by the "groups"
argument on the x-axis. Colors are determined by the \code{\link{colorGroup2}}
function.
}
\examples{
# No examples yet
}
