% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting.R
\name{qc_corr_hm}
\alias{qc_corr_hm}
\title{Correlation heatmap for QC report}
\usage{
qc_corr_hm(data, groups, sample_labels = NULL)
}
\arguments{
\item{data}{A data frame of intensity data, likely normalized. Rows should be
proteins and columns should be samples. Generally, a single element in the
normList slot of the object returned by \code{\link{process_data}}.}

\item{groups}{A character or factor vector, listing the group(s) the samples
belong to.}

\item{sample_labels}{Optional, a vector of sample labels to use. If not supplied,
defaults to using the column names in the data.}
}
\value{
A \code{\link[grid:gTree]{grid::gTree}} object of the ComplexHeatmap,
which can be plotted with \code{\link[grid:grid.draw]{grid::grid.draw}}.
}
\description{
Plots a ComplexHeatmap object showing the pairwise correlations of intensities
across samples. Samples are grouped by similarity on the x- and y-axis, with
labels colored by group and batch.
}
\examples{
# No examples yet
}
