% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting.R
\name{qc_dendrogram}
\alias{qc_dendrogram}
\title{Dendrogram for QC report}
\usage{
qc_dendrogram(
  data,
  groups = NULL,
  sampleLabels = NULL,
  top = 500,
  stdize = TRUE,
  clust.metric = "euclidean",
  clust.method = "complete",
  cex.names = 1,
  title = NULL,
  legend = TRUE
)
}
\arguments{
\item{data}{A data frame of intensity data, likely normalized. Rows should be
proteins and columns should be samples.}

\item{groups}{A character or factor vector, listing the group(s) the samples
belong to.}

\item{clust.metric}{The cluster metric used to define distance. Default is
"euclidean". See
\code{\link[ClassDiscovery:distanceMatrix]{ClassDiscovery::distanceMatrix}}
for options.}

\item{clust.method}{The agglomeration method to use for clustering. Default
is "complete", See \code{\link[stats:hclust]{stats::hclust}} for options.}

\item{cex.names}{Character expansion factor for sample labels. Default is 1.}
}
\value{
Invisibly returns a list with seven elements: \enumerate{
\item "hc"- The hclust object returned by
\code{\link[stats:hclust]{stats::hclust}}.
\item "dist"- The distance matrix returned by
\code{\link[ClassDiscovery:distanceMatrix]{ClassDiscovery::distanceMatrix}}.
\item "corr.data"- The data matrix that went into the correlation analysis.
\item "clust.method"- The clust.method argument used.
\item "clust.metric"- The clust.method argument used.
\item "stdize"- The stdize argument used.
\item "top"- The top argument used.
}
}
\description{
Performs and then plots a hierarchical clustering analysis of sample intensities
across samples. Sample labels are colored according to the "groups" argument,
with colors determined by the \code{\link{colorGroup2}} function. By default,
uses only the 500 most variable proteins for the analysis.
}
\examples{
# No examples yet
}
