% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_barplot_TotInt_by_Group.R
\name{qc_totInt_by_group}
\alias{qc_totInt_by_group}
\title{Generate Total Intensity Barplots by Group}
\usage{
qc_totInt_by_group(
  DAList,
  label_column,
  grouping_column,
  percentile,
  colors = NULL,
  nrow = NULL,
  ncol = NULL,
  legend.position = "none"
)
}
\arguments{
\item{DAList}{A list containing \code{data} (matrix) and \code{metadata} (data frame).}

\item{label_column}{A character string specifying the column name in metadata for sample labels.}

\item{grouping_column}{A character string specifying the column name in metadata for grouping.}

\item{percentile}{A numeric value (0-1) for the percentile threshold to flag low-intensity samples.}

\item{colors}{A vector of colors for groups. Defaults to a color-blind friendly palette (max 12, >12 uses Polychrome).}

\item{nrow}{Number of rows for facet wrapping.}

\item{ncol}{Number of columns for facet wrapping.}

\item{legend.position}{Position of the legend ('none', 'right', 'top', 'left', 'bottom').}
}
\value{
A list containing ggplot objects for total intensity, total number, and total missing values barplots.
}
\description{
This function creates barplots showing total intensity values for samples, grouped by a specified column.
It includes an optional percentile threshold to flag potential outliers and uses a color-blind friendly palette.
}
\examples{
# example code

\dontrun {

# example of barplot colored by group
barplot1 <- qc_totInt_by_group(DAList = results,
                            label_column = "sample",
                            grouping_column = "group",
                            percentile = 0,
                            colors = NULL,   # or c("#E69F00", "#56B4E9", "#009E73")
                            nrow = NULL,
                            ncol = NULL,
                            legend.position = "right")

}

}
