% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting.R
\name{qc_violin_plot}
\alias{qc_violin_plot}
\title{Violin plot for QC report}
\usage{
qc_violin_plot(data, groups = NULL, sample_labels = colnames(data))
}
\arguments{
\item{data}{A data frame of intensity data, likely normalized. Rows should be
proteins and columns should be samples.}

\item{groups}{A character or factor vector, listing the group(s) the samples
belong to.}

\item{sample_labels}{Optional, a vector of sample labels to use. If not supplied,
defaults to using the column names in the data.}
}
\value{
A ggplot object of the plot.
}
\description{
Makes a violin plot of per-sample intensities. Samples are grouped by the "groups"
argument on the x-axis.
}
\examples{
# No examples yet
}
