% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{read_DIA_data}
\alias{read_DIA_data}
\title{Load DIA data from a MaxQuant Samples Report}
\usage{
read_DIA_data(input_file = NULL, sample_IDs = NULL)
}
\arguments{
\item{input_file}{A MaxQuant Samples Report file to be imported.}

\item{sample_IDs}{Optional: a vector of sample IDs to import. Default is NULL, in which
case all sample IDs (determined from column names in the MaxQuant Samples Report) are
imported.}
}
\value{
A DAList containing data and annotation information.
}
\description{
THIS FUNCTION FOR UAMS INTERNAL USE. This function reads in and parses a UAMS-formatted
MaxQuant Samples Report file and creates a DAList object containing protein intensity data
and annotation data.
}
\examples{
\dontrun{
data <- read_DIA_data("path/to/Samples_Report.csv")
}

}
\keyword{UAMS}
