% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{read_DIA_data}
\alias{read_DIA_data}
\title{Loading DIA data from MaxQuant}
\usage{
read_DIA_data(input_file = NULL, sample_IDs = NULL)
}
\arguments{
\item{input_file}{The file of Maxquant data to be imported. Format and filetype
vary depending on the type of data to be analyzed. If not supplied, R will
ask you to pick a file from the file browser.}

\item{sample_IDs}{Optional: a vector of sample IDs to analyze. Default is NULL, in which
case all sample IDs are analyzed. These are determined from column names.}
}
\value{
A prototype of our new S3 list type (currently just a list).
}
\description{
A function to extract and load DIA MaxQuant data. The \code{read_DIA_data}
function is the main function, and it calls a number
of other subfunctions to process and prepare the data. See the links to
subfunctions for more info. Returns a list of data from maxquant, along with
some stats and other info. Also has some side effects: creates some log files.
}
\details{
Subfunctions (DIA): \code{\link{read_maxquant_delim}}, \code{\link{remove_contaminants}},
\code{\link{extract_protein_data}}, \code{\link{make_log}}.
}
\examples{
# No examples yet

}
