% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{read_maxquant_delim}
\alias{read_maxquant_delim}
\title{Import MaxQuant data}
\usage{
read_maxquant_delim(input_file)
}
\arguments{
\item{input_file}{A MaxQuant Samples Report file to be imported.}
}
\value{
A data frame, where each row is a (raw) protein, the first
few columns are annotation data, and the last columns are the individual
sample intensities (as character vectors).
}
\description{
THIS FUNCTION FOR UAMS INTERNAL USE. This is a subfunction of \code{\link{read_DIA_data}},
used to read in tabular data. This function is not meant to be called by users.
}
\keyword{UAMS,}
\keyword{internal}
