% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{read_maxquant_delim}
\alias{read_maxquant_delim}
\title{Import MaxQuant data}
\usage{
read_maxquant_delim(input_file)
}
\arguments{
\item{input_file}{The file of Maxquant data to be imported. Format and filetype
vary depending on the type of data to be analyzed. If not supplied, R will
ask you to pick a file from the file browser.}
}
\value{
a data frame, where each row is a (raw) protein, the first
few columns are protein ID data, and the last columns are the individual
sample intensities (as character vectors).
}
\description{
First subfunction called by \code{\link{read_DIA_data}}. Reads in the tabular
data and does some checks for required columns.
}
\examples{
# No examples yet
}
