% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_data_import.R
\name{remove_contaminants}
\alias{remove_contaminants}
\title{Remove protein contaminant rows}
\usage{
remove_contaminants(data, pipe, enrich)
}
\arguments{
\item{pipe}{Which analysis pipeline are you running? Options are "DIA", "TMT",
"phosphoTMT", and "LF". Output varies slightly based on the pupeline you
choose.}

\item{enrich}{Another aspect of pipeline? Options are "protein" and "phospho".
Seems semi-redundant?: DIA, TMT, and LF seem to always be "protein" in the
code. PhosphoTMT can be either protein or phospho, though that seems to vary
across subfunctions a little?}
}
\value{
A list of data with contaminants removed, with three elements:
(1) The filtered data frame, (2) the number of contaminant rows, and (3)
the number of rows retained.
}
\description{
The second subfunction called by \code{\link{extract_data}}. Removes
contaminant rows from the data.
}
\examples{
# No examples yet

}
