% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{remove_contaminants}
\alias{remove_contaminants}
\title{Remove protein contaminant rows}
\usage{
remove_contaminants(
  data,
  pipe = c("DIA", "TMT", "phosphoTMT", "LF"),
  enrich = c("protein", "phospho")
)
}
\arguments{
\item{data}{Raw Maxquant intensity data, from which contaminants need to be
removed. Within the \code{\link{extract_data}} function, this is "diaData",
which is the data object from the \code{\link{import_data}} function that has
gone through additional processing within the body of the
\code{\link{extract_data}} function}

\item{pipe}{Which analysis pipeline are you running? Options are "DIA", "TMT",
"phosphoTMT", and "LF". Output varies slightly based on the pipeline you
choose.}

\item{enrich}{Another aspect of pipeline? Options are "protein" and "phospho".
Seems semi-redundant?: DIA, TMT, and LF seem to always be "protein" in the
code. PhosphoTMT can be either protein or phospho, though that seems to vary
across subfunctions a little?}
}
\value{
A list of data with contaminants removed, with three elements:
\enumerate{
\item "data", a data frame with the contaminants removed. Rows are
proteins, columns include both protein info and individual intensities.
\item "no_contam", an integer giving the number of contaminant rows removed
\item "no_qfilter", an integer giving the number of retained rows
(ie, nrow() of the "data" slot).
}
}
\description{
The second subfunction called by \code{\link{extract_data}}. Removes
contaminant rows from the data.
}
\examples{
# No examples yet

}
