% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_analysis.R
\name{run_filtered_limma_analysis}
\alias{run_filtered_limma_analysis}
\title{Run limma analysis per contrast on filtered proteins}
\usage{
run_filtered_limma_analysis(
  DAList,
  design_formula = ~0 + group,
  pval_thresh = 0.05,
  lfc_thresh = 1,
  adj_method = "BH"
)
}
\arguments{
\item{DAList}{A DAList object containing full data and per-contrast filtered proteins.}

\item{design_formula}{A formula for the model design, e.g., ~0 + group.}

\item{pval_thresh}{P-value threshold used for defining significance. Default = 0.05.}

\item{lfc_thresh}{Log2 fold change threshold for significance. Default = 1.}

\item{adj_method}{Adjustment method for multiple testing ("BH", "BY", etc). Default = "BH".}
}
\value{
The input DAList, updated with filtered contrast-level results, eBayes fits,
and associated moving SDs and logFC z-scores.
}
\description{
This function subsets the DAList for each contrast according to
\code{filtered_proteins_per_contrast} and fits the limma model separately.
Useful when proteins are filtered differently per contrast.
}
\examples{
filtered_DAList <- run_filtered_limma_analysis(filtered_DAList)

}
