% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_report.R
\name{static_volcano_plot}
\alias{static_volcano_plot}
\title{Make a DE Volcano plot.}
\usage{
static_volcano_plot(data, lfc_thresh, pval_thresh, contrast, pval_type)
}
\arguments{
\item{data}{Per-contrast DE results to be plotted, as prepared by
\code{\link{prep_plot_model_data}}.}

\item{lfc_thresh}{The logFC threshold used to determine significance
(significant when |logFC| > lfc.tresh). LogFC are base 2.}

\item{pval_thresh}{The p-value threshold used to determine significance
(significant when p < pval_thresh).}

\item{contrast}{The contrast being plotted. Used for generating the plot title.}

\item{pval_type}{The type of p-value to plot. Can be "raw" or "adjusted".}
}
\value{
A ggplot object.
}
\description{
Internal function for plotting static versions of volcano plots.
}
\keyword{internal}
