% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_targets.R
\name{subset_targets}
\alias{subset_targets}
\title{Filter samples from targets file}
\usage{
subset_targets(targets, filter_column, rm.vals, ignore.case = TRUE)
}
\arguments{
\item{targets}{The input targets dataframe to be filtered}

\item{filter_column}{The column in which to check values.}

\item{rm.vals}{A character vector giving the values to search for in
filter_column. Rows matching these values will be fitlered out.}

\item{ignore.case}{Should the case of the strings in rm.vals be ignored? Is
TRUE by default, such that case does not have to match. If FALSE, case must
match for rows to be removed.}
}
\value{
A list with three elements:
\enumerate{
\item "targets", a dataframe of subsetted targets
\item "stats", a dataframe with statistics on filtering
\item "param", a dataframe giving the parameters used for filtering
}
}
\description{
Removes unwanted samples (rows) from the targets data frame This function is
basically \code{\link[base]{subset}} with extra functionality. It removes
rows if they contain certain values in a given column. Currently, can only
filter on one column at a time, but can provide multiple values to filter
against. By default, ignores the case of strings in rm.val. Calls
\code{\link{make_log}} as a subfunction.
}
\examples{
# No examples yet

}
