% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_contrasts.R
\name{validate_contrasts}
\alias{validate_contrasts}
\title{Validate contrasts format}
\usage{
validate_contrasts(contrast.tbl)
}
\arguments{
\item{contrast.tbl}{The imported data frame of contrasts.}
}
\value{
If all checks pass, invisible(TRUE).
}
\description{
Internal function that checks the format of the imported contrasts
and provides informative error messages about issues.
}
\examples{
# No examples yet
}
