% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class.R
\name{validate_DAList}
\alias{validate_DAList}
\title{DAList validator}
\usage{
validate_DAList(x)
}
\description{
Internal function for validating and normalizing a DAList-like object.
Tolerant of arbitrary result labels (custom contrast names).
Returns the (possibly re-ordered) object and ensures class "DAList".
}
\keyword{internal}
