% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class.R
\name{validate_DAList}
\alias{validate_DAList}
\title{DAList validator}
\usage{
validate_DAList(x)
}
\arguments{
\item{x}{An object to be tested if it is a valid DAList.}
}
\value{
If x was a valid DAList, returns x
}
\description{
Internal function for validating a DAList object
}
\keyword{internal}
