% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_report.R
\name{write_limma_plots}
\alias{write_limma_plots}
\title{Make interactive reports on DE}
\usage{
write_limma_plots(
  DIAlist = NULL,
  grouping_column = NULL,
  output_dir = NULL,
  tmp_subdir = "tmp",
  height = 1000,
  width = 1000
)
}
\arguments{
\item{DIAlist}{A DIAlist object, with statistical results.}

\item{grouping_column}{The name of the column in the metadata which
gives information on how to group samples for the interactive
expression plot.}

\item{output_dir}{The directory in which to create the reports and save the
plot files. No defaults, must be specified.}

\item{tmp_subdir}{The subdirectory within the output directory in which to
store temporary files. Deleted by default. Default is "tmp".}

\item{height}{The height of the interactive report objects, in pixels.
Default is 1000.}

\item{width}{The width of the interactive report objects, in pixels.
Default is 1000.}
}
\value{
Invisibly returns the input DIAlist.
}
\description{
A wrapper function that creates and saves interactive HTML reports summarizing
differential expression analyses for each contrast. Creates one HTML report
for each contrast. Also creates and saves a set of static .pdf plots. Current
behavior is to overwrite any previous reports or other files with the same name.
}
\examples{
# No examples yet
}
