% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_tables.R
\name{write_limma_results}
\alias{write_limma_results}
\title{Title}
\usage{
write_limma_results(
  model_results,
  annotation,
  ilab,
  out_dir = NULL,
  norm.method,
  enrich = c("protein", "phospho"),
  pipe = c("DIA", "TMT", "phosphoTMT", "LF"),
  overwrite = F,
  contrasts_subdir = NULL,
  summary_csv = NULL,
  combined_file_csv = NULL,
  BQ_csv = NULL,
  spreadsheet_xlsx = NULL,
  add_filter = T
)
}
\arguments{
\item{model_results}{TBD}

\item{annotation}{TBD}

\item{ilab}{TBD}

\item{out_dir}{TBD}

\item{norm.method}{TBD}

\item{enrich}{TBD}

\item{pipe}{TBD}

\item{overwrite}{TBD}

\item{contrasts_subdir}{TBD}

\item{summary_csv}{TBD}

\item{combined_file_csv}{TBD}

\item{BQ_csv}{TBD}

\item{spreadsheet_xlsx}{TBD}

\item{add_filter}{TBD}
}
\value{
TBD
}
\description{
TBD
}
\examples{
# No examples yet
}
