% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_tables.R
\name{write_limma_tables}
\alias{write_limma_tables}
\title{Write tables of limma results}
\usage{
write_limma_tables(
  DIAlist,
  ilab = NULL,
  output_dir = NULL,
  overwrite = F,
  contrasts_subdir = NULL,
  summary_csv = NULL,
  combined_file_csv = NULL,
  spreadsheet_xlsx = NULL,
  add_filter = T
)
}
\arguments{
\item{DIAlist}{A DIAlist object, with statistical results in the results slot.}

\item{ilab}{The ilab identifier for this project.}

\item{output_dir}{The directory in which to output tables. If not specified,
will construct a directory based on the ilab name and type of analysis.}

\item{overwrite}{Should results files be overwritten? Default is FALSE.}

\item{contrasts_subdir}{The subdirectory within output_dir to write the per-contrast
result .csv files. If not specified, will be "per_contrast_results".}

\item{summary_csv}{The filename of the csv file giving a summary of the number
of DE genes per contrast. If not specified, will be "DE_summary.csv".}

\item{combined_file_csv}{The filename of the combined results csv file. If
not specified, will be combined_results.csv.}

\item{spreadsheet_xlsx}{The filename of the Excel spreadsheet containing the
results. If not specified, will construct a name based on the ilab identifier.}

\item{add_filter}{Should per-column filters be added to the spreadsheet?
Default is TRUE. These can sometimes cause unstable spreadsheet files,
try setting to FALSE if you're having issues with the Excel output.}
}
\value{
Invisibly returns the input DIAlist.
}
\description{
This function prepares and then saves a number of tables of results:
\enumerate{
\item A .csv summarizing the number of direction of differentially
expressed genes for each contrast (see \code{summary_csv}).
\item An Excel spreadsheet of the intensity and
statistical results for all contrasts/terms in the results slot of the
DIAlist (see \code{spreadsheet_xlsx}).
\item A results .csv file with the same data as the Excel spreadsheet,
but in a flat .csv file withour formatting (see \code{combined_file_csv}).
\item A folder of results .csv file which contain the results for each
contrast.
}
}
\examples{
\dontrun{
  # Using defaults
  write_limma_tables(DIAlist)


 # Customize output directory
 # and filenames
 write_limma_tables(DIAlist,
                    output_dir = "DA_results",
                    contrasts_subdir = "by_contrast",
                    summary_csv = "my_summary.csv",
                    combined_file_csv = "all_results.csv",
                    spreadsheet_xlsx = "all_results.xlsx")

 # If the xlsx files have issues, try removing the
 # per-column filters in the spreadsheet
 write_limma_tables(DIAlist,
                    add_filter = F)
}
}
