% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_tables.R
\name{write_per_contrast_csvs}
\alias{write_per_contrast_csvs}
\title{Write per-contrast results csvs}
\usage{
write_per_contrast_csvs(
  annotation_df,
  data,
  results_statlist,
  output_dir,
  annotation_cols = NULL,
  metadata = NULL,
  group_col = "group",
  stat_cols = c("logFC", "P.Value", "adj.P.Val", "movingSD", "logFC_z_scores"),
  per_contrast_tags = NULL
)
}
\arguments{
\item{annotation_df}{A data frame of annotation data for each gene/protein.}

\item{data}{A data frame of average expression data for each sample.}

\item{results_statlist}{A list of per-contrast DE results.}

\item{output_dir}{The directory in which to save the per-contrast csv files.}

\item{annotation_cols}{Optional character vector of annotation column names to include.}

\item{metadata}{Optional data.frame of sample metadata (e.g. DAList$metadata).}

\item{group_col}{Character. Name of the grouping column within metadata to match contrast samples.}

\item{stat_cols}{Optional character vector of statistical columns to include from the results_statlist.}

\item{per_contrast_tags}{Optional list like DAList$tags$per_contrast; if present, will
be used to determine which sample groups/levels are involved in each contrast.}
}
\value{
A logical vector indicating whether each contrast file was successfully written.
}
\description{
Internal function used to write per-contrast statistical results to .csv files.
}
\keyword{internal}
