% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_tables.R
\name{write_per_contrast_csvs}
\alias{write_per_contrast_csvs}
\title{Write per-contrast results csvs}
\usage{
write_per_contrast_csvs(annotation_df, data, results_statlist, output_dir)
}
\arguments{
\item{annotation_df}{A data frame of annotation data for each gene/protein.}

\item{data}{A data frame of average expression data for each sample.}

\item{results_statlist}{A list of per-contrast DE results.}

\item{output_dir}{The directory in which to save the per-contrast csv files.}
}
\value{
A logical vector indicting whether each contrast file was
successfully written.
}
\description{
Internal function used to write per-contrast statistical results to a .csv
file.
}
\keyword{internal}
