% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_report.R
\name{write_proteinorm_report}
\alias{write_proteinorm_report}
\title{Create a normalization report}
\usage{
write_proteinorm_report(
  DIAlist,
  grouping_column = NULL,
  output_dir = NULL,
  filename = NULL,
  overwrite = FALSE,
  suppress_zoom_legend = FALSE
)
}
\arguments{
\item{DIAlist}{A DIAlist.}

\item{grouping_column}{The name of the column in the metadata which
gives information on how to group samples for normalization. Must be supplied:
some metrics can't be calculated for only one group.}

\item{output_dir}{The directory in which to save the report. If not provided,
will default to "protein_analysis/01_quality_control". If the directory does not
exist, it will be created.}

\item{filename}{The file name of the report to be saved. Must end in .pdf. Will
default to "proteiNorm_Report.pdf" if no file name is provided.}

\item{overwrite}{Should report file be overwritten if it already exists?
Default is FALSE.}

\item{suppress_zoom_legend}{Should the legend be removed from the zoomed
log2ratio plot? Default is FALSE.}
}
\value{
If report is created successfully, invisibly returns the input DIAlist.
}
\description{
Saves a PDF report containing a variety of plots which give
information about the performance of different normalization metrics.
The report is useful for choosing a normalization method to
use for downstream analysis.
}
\examples{
\dontrun{
# Group samples according to group identities
# in the "treatment" column of the metadata
write_proteinorm_report(DIAlist,
                        grouping_column = "treatment")

# Change the default directory and file names
write_proteinorm_report(DIAlist,
                        grouping_column = "treatment",
                        output_dir = "my/chosen/directory",
                        filename = "my_report.pdf")

# Overwrite an existing report
write_proteinorm_report(DIAlist,
                        grouping_column = "treatment",
                        overwrite = T)
}

}
